% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult_info.R
\docType{data}
\name{adult_info}
\alias{adult_info}
\title{DATASET_TITLE}
\format{A data frame with 455 rows and 4 variables:
\describe{
  \item{\code{variable_ine}}{character COLUMN_DESCRIPTION}
  \item{\code{descripcion_del_campo}}{character COLUMN_DESCRIPTION}
  \item{\code{tipo_variable}}{character COLUMN_DESCRIPTION}
  \item{\code{modulo}}{character COLUMN_DESCRIPTION}
}}
\usage{
adult_info
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
