% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adults_06_info.R
\docType{data}
\name{adults_06_info}
\alias{adults_06_info}
\title{adults_06_info}
\format{
A data frame with 619 rows and 6 variables:
\describe{
  \item{\code{campo}}{variable name}
  \item{\code{longitud}}{field length}
  \item{\code{posicion_inicio}}{start position in fixed-width file}
  \item{\code{posicion_final}}{end position in fixed-width file}
  \item{\code{descripcion_del_campo}}{variable description}
  \item{\code{valores_validos}}{valid values for the variable}
}
}
\usage{
adults_06_info
}
\description{
Design of the survey. Variable characteristics of adult survey ENSE 2006/07.
}
\keyword{datasets}
