% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adults_06_labels.R
\docType{data}
\name{adults_06_labels}
\alias{adults_06_labels}
\title{adults_06_labels}
\format{
A data frame with 2328 rows and 2 variables:
\describe{
  \item{\code{variable_ine}}{variable name}
  \item{\code{valores_ine}}{coded value}
}
}
\usage{
adults_06_labels
}
\description{
Value labels for the adult survey ENSE 2006/07.
}
\keyword{datasets}
