% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adults_12_labels.R
\docType{data}
\name{adults_12_labels}
\alias{adults_12_labels}
\title{DATASET_TITLE}
\format{
A data frame with 2714 rows and 3 variables:
\describe{
  \item{\code{valores_ine}}{character COLUMN_DESCRIPTION}
  \item{\code{valores}}{character COLUMN_DESCRIPTION}
  \item{\code{variable_ine}}{character COLUMN_DESCRIPTION}
}
}
\usage{
adults_12_labels
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
