% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adults_19.R
\docType{data}
\name{adults_19}
\alias{adults_19}
\title{adults_19}
\format{
A data frame with 23089 rows and 455 variables:
\describe{
  \item{\code{CCAA}}{}
  \item{\code{IDENTHOGAR}}{}
  \item{\code{A7_2a}}{}
  \item{\code{SEXOa}}{}
  \item{\code{EDADa}}{}
  \item{\code{ACTIVa}}{}
  \item{\code{PROXY_0}}{}
  \item{\code{PROXY_1}}{}
  \item{\code{PROXY_2}}{}
  \item{\code{PROXY_2b}}{}
  \item{\code{PROXY_3b}}{}
  \item{\code{PROXY_4}}{}
  \item{\code{PROXY_5}}{}
  \item{\code{E1_1}}{}
  \item{\code{E2_1a}}{}
  \item{\code{E2_1b}}{}
  \item{\code{E2_1c}}{}
  \item{\code{E2_1d}}{}
  \item{\code{E3}}{}
  \item{\code{E4}}{}
  \item{\code{E4b}}{}
  \item{\code{NIVEST}}{}
  \item{\code{F6}}{}
  \item{\code{F7}}{}
  \item{\code{F8_2}}{}
  \item{\code{F9_2}}{}
  \item{\code{F10}}{}
  \item{\code{F11}}{}
  \item{\code{F12}}{}
  \item{\code{F13}}{}
  \item{\code{F14a}}{}
  \item{\code{F14b}}{}
  \item{\code{F15}}{}
  \item{\code{F16}}{}
  \item{\code{F17}}{}
  \item{\code{F18a_2}}{}
  \item{\code{F18b_2}}{}
  \item{\code{F19a_2}}{}
  \item{\code{F19b_2}}{}
  \item{\code{F20}}{}
  \item{\code{G21}}{}
  \item{\code{G22}}{}
  \item{\code{G23}}{}
  \item{\code{G24}}{}
  \item{\code{G25a_1}}{}
  \item{\code{G25b_1}}{}
  \item{\code{G25c_1}}{}
  \item{\code{G25a_2}}{}
  \item{\code{G25b_2}}{}
  \item{\code{G25c_2}}{}
  \item{\code{G25a_3}}{}
  \item{\code{G25b_3}}{}
  \item{\code{G25c_3}}{}
  \item{\code{G25a_4}}{}
  \item{\code{G25b_4}}{}
  \item{\code{G25c_4}}{}
  \item{\code{G25a_5}}{}
  \item{\code{G25b_5}}{}
  \item{\code{G25c_5}}{}
  \item{\code{G25a_6}}{}
  \item{\code{G25b_6}}{}
  \item{\code{G25c_6}}{}
  \item{\code{G25a_7}}{}
  \item{\code{G25b_7}}{}
  \item{\code{G25c_7}}{}
  \item{\code{G25a_8}}{}
  \item{\code{G25b_8}}{}
  \item{\code{G25c_8}}{}
  \item{\code{G25a_9}}{}
  \item{\code{G25b_9}}{}
  \item{\code{G25c_9}}{}
  \item{\code{G25a_10}}{}
  \item{\code{G25b_10}}{}
  \item{\code{G25c_10}}{}
  \item{\code{G25a_11}}{}
  \item{\code{G25b_11}}{}
  \item{\code{G25c_11}}{}
  \item{\code{G25a_12}}{}
  \item{\code{G25b_12}}{}
  \item{\code{G25c_12}}{}
  \item{\code{G25a_13}}{}
  \item{\code{G25b_13}}{}
  \item{\code{G25c_13}}{}
  \item{\code{G25a_14}}{}
  \item{\code{G25b_14}}{}
  \item{\code{G25c_14}}{}
  \item{\code{G25a_15}}{}
  \item{\code{G25b_15}}{}
  \item{\code{G25c_15}}{}
  \item{\code{G25a_16}}{}
  \item{\code{G25b_16}}{}
  \item{\code{G25c_16}}{}
  \item{\code{G25a_17}}{}
  \item{\code{G25b_17}}{}
  \item{\code{G25c_17}}{}
  \item{\code{G25a_18}}{}
  \item{\code{G25b_18}}{}
  \item{\code{G25c_18}}{}
  \item{\code{G25a_19}}{}
  \item{\code{G25b_19}}{}
  \item{\code{G25c_19}}{}
  \item{\code{G25a_20}}{}
  \item{\code{G25b_20}}{}
  \item{\code{G25c_20}}{}
  \item{\code{G25a_21}}{}
  \item{\code{G25b_21}}{}
  \item{\code{G25c_21}}{}
  \item{\code{G25a_22}}{}
  \item{\code{G25b_22}}{}
  \item{\code{G25c_22}}{}
  \item{\code{G25a_23}}{}
  \item{\code{G25b_23}}{}
  \item{\code{G25c_23}}{}
  \item{\code{G25a_24}}{}
  \item{\code{G25b_24}}{}
  \item{\code{G25c_24}}{}
  \item{\code{G25a_25}}{}
  \item{\code{G25b_25}}{}
  \item{\code{G25c_25}}{}
  \item{\code{G25a_26}}{}
  \item{\code{G25b_26}}{}
  \item{\code{G25c_26}}{}
  \item{\code{G25a_27}}{}
  \item{\code{G25b_27}}{}
  \item{\code{G25c_27}}{}
  \item{\code{G25a_28}}{}
  \item{\code{G25b_28}}{}
  \item{\code{G25c_28}}{}
  \item{\code{G25a_29}}{}
  \item{\code{G25b_29}}{}
  \item{\code{G25c_29}}{}
  \item{\code{G25a_30}}{}
  \item{\code{G25b_30}}{}
  \item{\code{G25c_30}}{}
  \item{\code{G25a_31}}{}
  \item{\code{G25b_31}}{}
  \item{\code{G25c_31}}{}
  \item{\code{G25a_32}}{}
  \item{\code{G25b_32}}{}
  \item{\code{G25c_32}}{}
  \item{\code{H26_1}}{}
  \item{\code{H26_2}}{}
  \item{\code{H26_3}}{}
  \item{\code{H27}}{}
  \item{\code{I28_1}}{}
  \item{\code{I28_2}}{}
  \item{\code{I29_1}}{}
  \item{\code{I29_2}}{}
  \item{\code{K32}}{}
  \item{\code{K33}}{}
  \item{\code{K34}}{}
  \item{\code{K35}}{}
  \item{\code{K36}}{}
  \item{\code{K37}}{}
  \item{\code{K38}}{}
  \item{\code{K38a}}{}
  \item{\code{L39_1}}{}
  \item{\code{L39_2}}{}
  \item{\code{L39_3}}{}
  \item{\code{L39_4}}{}
  \item{\code{L39_5}}{}
  \item{\code{L40}}{}
  \item{\code{L41}}{}
  \item{\code{L42_1}}{}
  \item{\code{L42_2}}{}
  \item{\code{L42_3}}{}
  \item{\code{L42_4}}{}
  \item{\code{L42_5}}{}
  \item{\code{L42_6}}{}
  \item{\code{L42_7}}{}
  \item{\code{L43}}{}
  \item{\code{L44}}{}
  \item{\code{L45}}{}
  \item{\code{L46}}{}
  \item{\code{M47_1}}{}
  \item{\code{M47_2}}{}
  \item{\code{M47_3}}{}
  \item{\code{M47_4}}{}
  \item{\code{M47_5}}{}
  \item{\code{M47_6}}{}
  \item{\code{M47_7}}{}
  \item{\code{M47_8}}{}
  \item{\code{M47_9}}{}
  \item{\code{M47_10}}{}
  \item{\code{M47_11}}{}
  \item{\code{M47_12}}{}
  \item{\code{M47a}}{}
  \item{\code{M47b}}{}
  \item{\code{N48}}{}
  \item{\code{N49}}{}
  \item{\code{N50}}{}
  \item{\code{N51}}{}
  \item{\code{N52}}{}
  \item{\code{N53}}{}
  \item{\code{N54}}{}
  \item{\code{N55_1}}{}
  \item{\code{N55_2}}{}
  \item{\code{N55_3}}{}
  \item{\code{N56_1}}{}
  \item{\code{N56_2}}{}
  \item{\code{N56_3}}{}
  \item{\code{N57}}{}
  \item{\code{N58_1}}{}
  \item{\code{N58_2}}{}
  \item{\code{N58_3}}{}
  \item{\code{N59}}{}
  \item{\code{N60_1}}{}
  \item{\code{N60_2}}{}
  \item{\code{N60_3}}{}
  \item{\code{N60_4}}{}
  \item{\code{N60a_1}}{}
  \item{\code{N60a_2}}{}
  \item{\code{N60a_3}}{}
  \item{\code{N60a_4}}{}
  \item{\code{N61_1}}{}
  \item{\code{N61_2}}{}
  \item{\code{N61_3}}{}
  \item{\code{N61_4}}{}
  \item{\code{N61_5}}{}
  \item{\code{N62}}{}
  \item{\code{N62b}}{}
  \item{\code{N63_1}}{}
  \item{\code{N63_2}}{}
  \item{\code{N63_3}}{}
  \item{\code{N63_4}}{}
  \item{\code{N63_5}}{}
  \item{\code{N63_6}}{}
  \item{\code{N63_7}}{}
  \item{\code{N63_8}}{}
  \item{\code{N63_9}}{}
  \item{\code{N63_10}}{}
  \item{\code{N64}}{}
  \item{\code{N65_1}}{}
  \item{\code{N65_2}}{}
  \item{\code{N65_3}}{}
  \item{\code{N65_4}}{}
  \item{\code{N65_5}}{}
  \item{\code{N65_6}}{}
  \item{\code{N65_7}}{}
  \item{\code{N65_8}}{}
  \item{\code{O66}}{}
  \item{\code{O67}}{}
  \item{\code{O69}}{}
  \item{\code{O70}}{}
  \item{\code{O71}}{}
  \item{\code{O72}}{}
  \item{\code{O73}}{}
  \item{\code{O74}}{}
  \item{\code{O75}}{}
  \item{\code{O76}}{}
  \item{\code{O77}}{}
  \item{\code{O78}}{}
  \item{\code{O79}}{}
  \item{\code{O80_1}}{}
  \item{\code{O80_2}}{}
  \item{\code{O80_3}}{}
  \item{\code{O81_1}}{}
  \item{\code{O81_2}}{}
  \item{\code{O81_3}}{}
  \item{\code{O82_1}}{}
  \item{\code{O82_2}}{}
  \item{\code{O83}}{}
  \item{\code{O84_1}}{}
  \item{\code{O84_2}}{}
  \item{\code{O84_3}}{}
  \item{\code{O84_4}}{}
  \item{\code{O84_5}}{}
  \item{\code{O84_6}}{}
  \item{\code{O84_7}}{}
  \item{\code{O84_8}}{}
  \item{\code{O84_9}}{}
  \item{\code{P85}}{}
  \item{\code{P86}}{}
  \item{\code{P87_1a}}{}
  \item{\code{P87_1b}}{}
  \item{\code{P87_2a}}{}
  \item{\code{P87_2b}}{}
  \item{\code{P87_3a}}{}
  \item{\code{P87_3b}}{}
  \item{\code{P87_4a}}{}
  \item{\code{P87_4b}}{}
  \item{\code{P87_5a}}{}
  \item{\code{P87_5b}}{}
  \item{\code{P87_6a}}{}
  \item{\code{P87_6b}}{}
  \item{\code{P87_7a}}{}
  \item{\code{P87_7b}}{}
  \item{\code{P87_8a}}{}
  \item{\code{P87_8b}}{}
  \item{\code{P87_9a}}{}
  \item{\code{P87_9b}}{}
  \item{\code{P87_10a}}{}
  \item{\code{P87_10b}}{}
  \item{\code{P87_11a}}{}
  \item{\code{P87_11b}}{}
  \item{\code{P87_12a}}{}
  \item{\code{P87_12b}}{}
  \item{\code{P87_13a}}{}
  \item{\code{P87_13b}}{}
  \item{\code{P87_14a}}{}
  \item{\code{P87_14b}}{}
  \item{\code{P87_15a}}{}
  \item{\code{P87_15b}}{}
  \item{\code{P87_16a}}{}
  \item{\code{P87_16b}}{}
  \item{\code{P87_17a}}{}
  \item{\code{P87_17b}}{}
  \item{\code{P87_18a}}{}
  \item{\code{P87_18b}}{}
  \item{\code{P87_19a}}{}
  \item{\code{P87_19b}}{}
  \item{\code{P87_20a}}{}
  \item{\code{P87_20b}}{}
  \item{\code{P87_21a}}{}
  \item{\code{P87_21b}}{}
  \item{\code{P87_22a}}{}
  \item{\code{P87_22b}}{}
  \item{\code{P87_23a}}{}
  \item{\code{P87_23b}}{}
  \item{\code{Q88}}{}
  \item{\code{Q89}}{}
  \item{\code{Q90}}{}
  \item{\code{Q91}}{}
  \item{\code{Q92}}{}
  \item{\code{Q93}}{}
  \item{\code{Q94}}{}
  \item{\code{Q95}}{}
  \item{\code{Q96}}{}
  \item{\code{Q97}}{}
  \item{\code{Q98}}{}
  \item{\code{Q99}}{}
  \item{\code{Q100}}{}
  \item{\code{Q101}}{}
  \item{\code{Q102}}{}
  \item{\code{Q103}}{}
  \item{\code{Q104}}{}
  \item{\code{Q105}}{}
  \item{\code{R106}}{}
  \item{\code{R107}}{}
  \item{\code{R108_1}}{}
  \item{\code{R108_2}}{}
  \item{\code{R108_3}}{}
  \item{\code{R108_4}}{}
  \item{\code{S109}}{}
  \item{\code{S110}}{}
  \item{\code{T111}}{}
  \item{\code{T112}}{}
  \item{\code{T113}}{}
  \item{\code{T114_1}}{}
  \item{\code{T114_2}}{}
  \item{\code{T115}}{}
  \item{\code{T116_1}}{}
  \item{\code{T116_2}}{}
  \item{\code{T117}}{}
  \item{\code{T118_1}}{}
  \item{\code{T118_2}}{}
  \item{\code{T119_1}}{}
  \item{\code{T119_2}}{}
  \item{\code{U120_1}}{}
  \item{\code{U120_1a}}{}
  \item{\code{U120_2}}{}
  \item{\code{U120_3}}{}
  \item{\code{U120_4}}{}
  \item{\code{U120_5}}{}
  \item{\code{U120_6}}{}
  \item{\code{U120_7}}{}
  \item{\code{U120_7a}}{}
  \item{\code{U120_8}}{}
  \item{\code{U120_9}}{}
  \item{\code{U120_10}}{}
  \item{\code{U120_11}}{}
  \item{\code{U120_12}}{}
  \item{\code{U120_13}}{}
  \item{\code{U120_14}}{}
  \item{\code{U120_15}}{}
  \item{\code{U120_15a}}{}
  \item{\code{U120FZ}}{}
  \item{\code{U120CANTFZ}}{}
  \item{\code{U2_120F}}{}
  \item{\code{V121}}{}
  \item{\code{V122}}{}
  \item{\code{V123}}{}
  \item{\code{V124}}{}
  \item{\code{V125}}{}
  \item{\code{V126}}{}
  \item{\code{W127}}{}
  \item{\code{W128Cer}}{}
  \item{\code{W128Cer_1}}{}
  \item{\code{W128Cer_2}}{}
  \item{\code{W128Cer_3}}{}
  \item{\code{W128Cer_4}}{}
  \item{\code{W128Cer_5}}{}
  \item{\code{W128Cer_6}}{}
  \item{\code{W128Cer_7}}{}
  \item{\code{W128Vin}}{}
  \item{\code{W128Vin_1}}{}
  \item{\code{W128Vin_2}}{}
  \item{\code{W128Vin_3}}{}
  \item{\code{W128Vin_4}}{}
  \item{\code{W128Vin_5}}{}
  \item{\code{W128Vin_6}}{}
  \item{\code{W128Vin_7}}{}
  \item{\code{W128Vermut}}{}
  \item{\code{W128Vermut_1}}{}
  \item{\code{W128Vermut_2}}{}
  \item{\code{W128Vermut_3}}{}
  \item{\code{W128Vermut_4}}{}
  \item{\code{W128Vermut_5}}{}
  \item{\code{W128Vermut_6}}{}
  \item{\code{W128Vermut_7}}{}
  \item{\code{W128Lic}}{}
  \item{\code{W128Lic_1}}{}
  \item{\code{W128Lic_2}}{}
  \item{\code{W128Lic_3}}{}
  \item{\code{W128Lic_4}}{}
  \item{\code{W128Lic_5}}{}
  \item{\code{W128Lic_6}}{}
  \item{\code{W128Lic_7}}{}
  \item{\code{W128Comb}}{}
  \item{\code{W128Comb_1}}{}
  \item{\code{W128Comb_2}}{}
  \item{\code{W128Comb_3}}{}
  \item{\code{W128Comb_4}}{}
  \item{\code{W128Comb_5}}{}
  \item{\code{W128Comb_6}}{}
  \item{\code{W128Comb_7}}{}
  \item{\code{W128Sidra}}{}
  \item{\code{W128Sidra_1}}{}
  \item{\code{W128Sidra_2}}{}
  \item{\code{W128Sidra_3}}{}
  \item{\code{W128Sidra_4}}{}
  \item{\code{W128Sidra_5}}{}
  \item{\code{W128Sidra_6}}{}
  \item{\code{W128Sidra_7}}{}
  \item{\code{W129}}{}
  \item{\code{X130_1}}{}
  \item{\code{X130_2}}{}
  \item{\code{X130_3}}{}
  \item{\code{X130_4}}{}
  \item{\code{X130_5}}{}
  \item{\code{X130_6}}{}
  \item{\code{X130_7}}{}
  \item{\code{X130_8}}{}
  \item{\code{X130_9}}{}
  \item{\code{X130_10}}{}
  \item{\code{X130_11}}{}
  \item{\code{Y133}}{}
  \item{\code{Y134}}{}
  \item{\code{Y135}}{}
  \item{\code{FACTORADULTO}}{}
  \item{\code{CLASE_PR}}{}
  \item{\code{IMCa}}{}
  \item{\code{CMD1}}{}
  \item{\code{CMD2}}{}
  \item{\code{CMD3}}{}
}
}
\usage{
adults_19
}
\description{
Adult survey of the ENSE 2017/19.
}
\keyword{datasets}
