% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adults_19_labels.R
\docType{data}
\name{adults_19_labels}
\alias{adults_19_labels}
\title{adults_19_labels}
\format{
A data frame with 2305 rows and 3 variables:
\describe{
  \item{\code{valores_ine}}{values of the variable}
  \item{\code{valores}}{labels}
  \item{\code{variable_ine}}{name of the variable}
}
}
\usage{
adults_19_labels
}
\description{
Provides labels to the values of the adults survey. ENSE 17/19
}
\details{
DETAILS
}
\keyword{datasets}
