% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adults_info_final.R
\docType{data}
\name{adults_info_final}
\alias{adults_info_final}
\title{DATASET_TITLE}
\format{A data frame with 455 rows and 7 variables:
\describe{
  \item{\code{variable_ine}}{character COLUMN_DESCRIPTION}
  \item{\code{longitud}}{double COLUMN_DESCRIPTION}
  \item{\code{posicion_inicio}}{double COLUMN_DESCRIPTION}
  \item{\code{posicion_final}}{double COLUMN_DESCRIPTION}
  \item{\code{descripcion_del_campo}}{character COLUMN_DESCRIPTION}
  \item{\code{tipo_variable}}{character COLUMN_DESCRIPTION}
  \item{\code{modulo}}{logical COLUMN_DESCRIPTION}
}}
\usage{
adults_info_final
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
