% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/children_06.R
\docType{data}
\name{children_06}
\alias{children_06}
\title{children_06}
\format{
A data frame with 9122 rows and 350 variables:
\describe{
  \item{\code{NIVEL}}{}
  \item{\code{CCAA}}{}
  \item{\code{PROV}}{}
  \item{\code{ESTRATO}}{}
  \item{\code{SECCION}}{}
  \item{\code{SECC_DC}}{}
  \item{\code{NIDENTIF}}{}
  \item{\code{AÑO}}{}
  \item{\code{TRIM}}{}
  \item{\code{NORDEN}}{}
  \item{\code{EDAD}}{}
  \item{\code{EDADMES}}{}
  \item{\code{SEXO}}{}
  \item{\code{P3_3}}{}
  \item{\code{P3_4}}{}
  \item{\code{P3_5}}{}
  \item{\code{P4_1}}{}
  \item{\code{P4_0_1}}{}
  \item{\code{P4_0_2}}{}
  \item{\code{P4_3}}{}
  \item{\code{P4_4}}{}
  \item{\code{A1}}{}
  \item{\code{A2_1A}}{}
  \item{\code{A2_1B}}{}
  \item{\code{A2_1C}}{}
  \item{\code{A2_1D}}{}
  \item{\code{A2_2A}}{}
  \item{\code{A2_2B}}{}
  \item{\code{A2_2C}}{}
  \item{\code{A2_2D}}{}
  \item{\code{A2_3A}}{}
  \item{\code{A2_3B}}{}
  \item{\code{A2_3C}}{}
  \item{\code{A2_3D}}{}
  \item{\code{A2_4A}}{}
  \item{\code{A2_4B}}{}
  \item{\code{A2_4C}}{}
  \item{\code{A2_4D}}{}
  \item{\code{A2_5A}}{}
  \item{\code{A2_5B}}{}
  \item{\code{A2_5C}}{}
  \item{\code{A2_5D}}{}
  \item{\code{A2_6A}}{}
  \item{\code{A2_6B}}{}
  \item{\code{A2_6C}}{}
  \item{\code{A2_6D}}{}
  \item{\code{A2_7A}}{}
  \item{\code{A2_7B}}{}
  \item{\code{A2_7C}}{}
  \item{\code{A2_7D}}{}
  \item{\code{A2_8A}}{}
  \item{\code{A2_8LIT}}{}
  \item{\code{A2_8B}}{}
  \item{\code{A2_8C}}{}
  \item{\code{A2_8D}}{}
  \item{\code{A2_9A}}{}
  \item{\code{A2_9LIT}}{}
  \item{\code{A2_9B}}{}
  \item{\code{A2_9C}}{}
  \item{\code{A2_9D}}{}
  \item{\code{A3}}{}
  \item{\code{B4}}{}
  \item{\code{B5_1A}}{}
  \item{\code{B5_1B}}{}
  \item{\code{B5_2A}}{}
  \item{\code{B5_2B}}{}
  \item{\code{B5_3A}}{}
  \item{\code{B5_3B}}{}
  \item{\code{B5_4A}}{}
  \item{\code{B5_4B}}{}
  \item{\code{B5_5A}}{}
  \item{\code{B5_5B}}{}
  \item{\code{B5_6A}}{}
  \item{\code{B5_6B}}{}
  \item{\code{B5_7A}}{}
  \item{\code{B5_7B}}{}
  \item{\code{B5_8A}}{}
  \item{\code{B5_8B}}{}
  \item{\code{B6}}{}
  \item{\code{B7}}{}
  \item{\code{B8_1}}{}
  \item{\code{B8_2}}{}
  \item{\code{B8_3}}{}
  \item{\code{B8_4}}{}
  \item{\code{B8_5}}{}
  \item{\code{B9}}{}
  \item{\code{B10}}{}
  \item{\code{C11_1}}{}
  \item{\code{C11_2}}{}
  \item{\code{C12_1}}{}
  \item{\code{C12_2}}{}
  \item{\code{C13_1A}}{}
  \item{\code{C13_1B}}{}
  \item{\code{C13_2A}}{}
  \item{\code{C13_2B}}{}
  \item{\code{C13_3A}}{}
  \item{\code{C13_3B}}{}
  \item{\code{C13_4A}}{}
  \item{\code{C13_4B}}{}
  \item{\code{C13_5A}}{}
  \item{\code{C13_5B}}{}
  \item{\code{C13_6A}}{}
  \item{\code{C13_6B}}{}
  \item{\code{C13_7A}}{}
  \item{\code{C13_7B}}{}
  \item{\code{C13_8A}}{}
  \item{\code{C13_8B}}{}
  \item{\code{C13_9A}}{}
  \item{\code{C13_9B}}{}
  \item{\code{C13_10A}}{}
  \item{\code{C13_10B}}{}
  \item{\code{C13_11A}}{}
  \item{\code{C13_11B}}{}
  \item{\code{C13_12A}}{}
  \item{\code{C13_12B}}{}
  \item{\code{C13_13A}}{}
  \item{\code{C13_13B}}{}
  \item{\code{C13_14A}}{}
  \item{\code{C13_14B}}{}
  \item{\code{D14_1A}}{}
  \item{\code{D14_1B}}{}
  \item{\code{D14_2A}}{}
  \item{\code{D14_2B}}{}
  \item{\code{D14_3A}}{}
  \item{\code{D14_3B}}{}
  \item{\code{D14_4A}}{}
  \item{\code{D14_4B}}{}
  \item{\code{D14_5A}}{}
  \item{\code{D14_5B}}{}
  \item{\code{D14_6A}}{}
  \item{\code{D14_6B}}{}
  \item{\code{D14_7A}}{}
  \item{\code{D14_7B}}{}
  \item{\code{D14_8A}}{}
  \item{\code{D14_8B}}{}
  \item{\code{D14_9A}}{}
  \item{\code{D14_9B}}{}
  \item{\code{D14_10A}}{}
  \item{\code{D14_10B}}{}
  \item{\code{D14_11A}}{}
  \item{\code{D14_11B}}{}
  \item{\code{D14_12A}}{}
  \item{\code{D14_12B}}{}
  \item{\code{D14_13A}}{}
  \item{\code{D14_13B}}{}
  \item{\code{D14_14A}}{}
  \item{\code{D14_15A}}{}
  \item{\code{E15_0}}{}
  \item{\code{E15_1}}{}
  \item{\code{E15_2}}{}
  \item{\code{E15_3}}{}
  \item{\code{E15_4}}{}
  \item{\code{E15_5}}{}
  \item{\code{E15_6}}{}
  \item{\code{E15_7}}{}
  \item{\code{E15_8}}{}
  \item{\code{E15_9}}{}
  \item{\code{E15_10}}{}
  \item{\code{E15_11}}{}
  \item{\code{E15_12}}{}
  \item{\code{E15_13}}{}
  \item{\code{E15_14}}{}
  \item{\code{E15_15}}{}
  \item{\code{E15_16}}{}
  \item{\code{E15_17}}{}
  \item{\code{E15_18}}{}
  \item{\code{E15_19}}{}
  \item{\code{E15_20}}{}
  \item{\code{E15_21}}{}
  \item{\code{E15_22}}{}
  \item{\code{E15_23}}{}
  \item{\code{E15_24}}{}
  \item{\code{E15_25}}{}
  \item{\code{F1_16_1}}{}
  \item{\code{F1_16_2}}{}
  \item{\code{F1_16_3}}{}
  \item{\code{F1_17}}{}
  \item{\code{F1_18}}{}
  \item{\code{F1_19}}{}
  \item{\code{F1_20}}{}
  \item{\code{F1_21}}{}
  \item{\code{F1_22_1}}{}
  \item{\code{F1_22_2}}{}
  \item{\code{F1_22_3}}{}
  \item{\code{F1_23_1}}{}
  \item{\code{F1_23_2}}{}
  \item{\code{F1_23_3}}{}
  \item{\code{F1_24}}{}
  \item{\code{F1_25}}{}
  \item{\code{F1_26_1A}}{}
  \item{\code{F1_26_1B}}{}
  \item{\code{F1_26_1C}}{}
  \item{\code{F1_26_2A}}{}
  \item{\code{F1_26_2B}}{}
  \item{\code{F1_26_2C}}{}
  \item{\code{F1_26_3A}}{}
  \item{\code{F1_26_3B}}{}
  \item{\code{F1_26_3C}}{}
  \item{\code{F1_26_4A}}{}
  \item{\code{F1_26_4B}}{}
  \item{\code{F1_26_4C}}{}
  \item{\code{F1_26_5A}}{}
  \item{\code{F1_26_5B}}{}
  \item{\code{F1_26_5C}}{}
  \item{\code{F1_26_6A}}{}
  \item{\code{F1_26_6B}}{}
  \item{\code{F1_26_6C}}{}
  \item{\code{F1_27_1A}}{}
  \item{\code{F1_27_1B}}{}
  \item{\code{F1_27_2A}}{}
  \item{\code{F1_27_2B}}{}
  \item{\code{F2_28}}{}
  \item{\code{F2_29}}{}
  \item{\code{F3_30_1}}{}
  \item{\code{F3_30_2}}{}
  \item{\code{F3_30_3}}{}
  \item{\code{F3_30_4}}{}
  \item{\code{F3_30_5}}{}
  \item{\code{F3_30_6}}{}
  \item{\code{F3_30_7}}{}
  \item{\code{F3_30_8}}{}
  \item{\code{F4_31_1}}{}
  \item{\code{F4_31_2}}{}
  \item{\code{F4_31_3}}{}
  \item{\code{F4_31_4}}{}
  \item{\code{F4_32_1}}{}
  \item{\code{F4_32_2}}{}
  \item{\code{F4_32_3}}{}
  \item{\code{F4_32_4}}{}
  \item{\code{F4_32_5}}{}
  \item{\code{F4_32_6}}{}
  \item{\code{F4_32_7}}{}
  \item{\code{F4_32_8}}{}
  \item{\code{F4_32_9}}{}
  \item{\code{F4_33}}{}
  \item{\code{F4_34}}{}
  \item{\code{F4_35}}{}
  \item{\code{F4_36_1}}{}
  \item{\code{F4_36_2}}{}
  \item{\code{F4_36_3}}{}
  \item{\code{F4_36_4}}{}
  \item{\code{F4_36_5}}{}
  \item{\code{F5_37}}{}
  \item{\code{F5_38}}{}
  \item{\code{F5_39}}{}
  \item{\code{F5_40}}{}
  \item{\code{F5_41}}{}
  \item{\code{F5_42}}{}
  \item{\code{F5_43}}{}
  \item{\code{F5_44}}{}
  \item{\code{F5_45}}{}
  \item{\code{F5_46}}{}
  \item{\code{F5_47}}{}
  \item{\code{F5_48}}{}
  \item{\code{F6_49}}{}
  \item{\code{F6_50}}{}
  \item{\code{F6_51_1}}{}
  \item{\code{F6_51_2}}{}
  \item{\code{F6_51_3}}{}
  \item{\code{F6_52_1}}{}
  \item{\code{F6_52_2}}{}
  \item{\code{F6_52_3}}{}
  \item{\code{F6_53_1}}{}
  \item{\code{F6_53_2}}{}
  \item{\code{F6_54}}{}
  \item{\code{F6_55}}{}
  \item{\code{G1_56}}{}
  \item{\code{G1_57}}{}
  \item{\code{G1_58_0}}{}
  \item{\code{G1_58}}{}
  \item{\code{G1_59_1A}}{}
  \item{\code{G1_59_1B}}{}
  \item{\code{G1_59_2A}}{}
  \item{\code{G1_59_2B}}{}
  \item{\code{G1_60}}{}
  \item{\code{G1_61_1A}}{}
  \item{\code{G1_61_1B}}{}
  \item{\code{G1_61_2A}}{}
  \item{\code{G1_61_2B}}{}
  \item{\code{G1_62_0}}{}
  \item{\code{G2_63}}{}
  \item{\code{G2_64_1}}{}
  \item{\code{G2_64_2}}{}
  \item{\code{G2_65}}{}
  \item{\code{G2_66_1}}{}
  \item{\code{G2_66_2}}{}
  \item{\code{G2_67}}{}
  \item{\code{G2_68_1}}{}
  \item{\code{G2_68_2}}{}
  \item{\code{G1_69}}{}
  \item{\code{G2_70_1}}{}
  \item{\code{G2_70_2}}{}
  \item{\code{G2_70_3}}{}
  \item{\code{G2_70_4}}{}
  \item{\code{G2_70_5}}{}
  \item{\code{G2_70_6}}{}
  \item{\code{G2_71_1}}{}
  \item{\code{G2_71_2}}{}
  \item{\code{G2_71_3}}{}
  \item{\code{G2_71_4}}{}
  \item{\code{G2_71_5}}{}
  \item{\code{G2_71_6}}{}
  \item{\code{G2_71_7}}{}
  \item{\code{G2_71_8}}{}
  \item{\code{G2_71_9}}{}
  \item{\code{G2_71_10}}{}
  \item{\code{G2_71_11}}{}
  \item{\code{G2_71_12}}{}
  \item{\code{G2_71_13}}{}
  \item{\code{G2_71_14}}{}
  \item{\code{G2_72}}{}
  \item{\code{G2_73}}{}
  \item{\code{H74_0}}{}
  \item{\code{H75}}{}
  \item{\code{H76}}{}
  \item{\code{H77_1}}{}
  \item{\code{H77_2}}{}
  \item{\code{H78}}{}
  \item{\code{H79}}{}
  \item{\code{I80}}{}
  \item{\code{I81_1A}}{}
  \item{\code{I81_1B}}{}
  \item{\code{I81_1C}}{}
  \item{\code{I81_1D}}{}
  \item{\code{I81_1E}}{}
  \item{\code{I81_2A}}{}
  \item{\code{I81_2B}}{}
  \item{\code{I81_2C}}{}
  \item{\code{I81_2D}}{}
  \item{\code{I81_2E}}{}
  \item{\code{I81_3A}}{}
  \item{\code{I81_3B}}{}
  \item{\code{I81_3C}}{}
  \item{\code{I81_3D}}{}
  \item{\code{I81_3E}}{}
  \item{\code{K93}}{}
  \item{\code{K94}}{}
  \item{\code{K95}}{}
  \item{\code{K96}}{}
  \item{\code{K97}}{}
  \item{\code{FACTOR}}{}
  \item{\code{TMUNI}}{}
  \item{\code{SPCLASE}}{}
  \item{\code{MSCALVIDA}}{}
  \item{\code{MSSALMENT1}}{}
  \item{\code{MSSALMENT2}}{}
  \item{\code{MSSALMENT3}}{}
  \item{\code{MSSALMENT4}}{}
  \item{\code{MSSALMENT5}}{}
  \item{\code{MSIMC}}{}
}
}
\usage{
children_06
}
\description{
Children survey of the ENSE 2006/07.
}
\keyword{datasets}
