% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/children_06_labels.R
\docType{data}
\name{children_06_labels}
\alias{children_06_labels}
\title{children_06_labels}
\format{
A data frame with 1306 rows and 2 variables:
\describe{
  \item{\code{variable_ine}}{Name of the variable}
  \item{\code{valores_ine}}{Values and labels of the variable}
}
}
\usage{
children_06_labels
}
\description{
Provides labels to the values of the children survey. ENSE 2006/07
}
\details{
DETAILS
}
\keyword{datasets}
