% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/children_12_info.R
\docType{data}
\name{children_12_info}
\alias{children_12_info}
\title{children_12_info}
\format{
A data frame with 271 rows and 6 variables:
\describe{
  \item{\code{campo}}{Name of the variable}
  \item{\code{longitud}}{Length of the variable. Number of possible characters it may take.}
  \item{\code{posicion_inicio}}{Initial position in the text file. Information for parsing microdata}
  \item{\code{posicion_final}}{Final position in the text file. Information for parsing microdata}
  \item{\code{descripcion_del_campo}}{Description of the variable (in Spanish)}
  \item{\code{tipo_variable}}{Possible values the variable may take}
}
}
\usage{
children_12_info
}
\description{
Design of the survey. Variable characteristics of children survey. ENSE 2011/12
}
\details{
DETAILS
}
\keyword{datasets}
