% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/children_12_labels.R
\docType{data}
\name{children_12_labels}
\alias{children_12_labels}
\title{children_12_labels}
\format{
A data frame with 1289 rows and 3 variables:
\describe{
  \item{\code{valores_ine}}{values in survey}
  \item{\code{valores}}{labels of values}
  \item{\code{variable_ine}}{name of the variable}
}
}
\usage{
children_12_labels
}
\description{
Provides labels to the values of the household survey. ENSE 11/12
}
\details{
DETAILS
}
\keyword{datasets}
