% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/children_19.R
\docType{data}
\name{children_19}
\alias{children_19}
\title{children_19}
\format{
A data frame with 6106 rows and 267 variables:
\describe{
  \item{\code{CCAA}}{}
  \item{\code{IDENTHOGAR}}{}
  \item{\code{A7_2m}}{}
  \item{\code{SEXOm}}{}
  \item{\code{EDADm}}{}
  \item{\code{Informante_1}}{}
  \item{\code{Informante_1a}}{}
  \item{\code{Informante_3}}{}
  \item{\code{Informante_4}}{}
  \item{\code{Informante_5}}{}
  \item{\code{A1_1}}{}
  \item{\code{A2_1a}}{}
  \item{\code{A2_1b}}{}
  \item{\code{A2_1c}}{}
  \item{\code{A2_1d}}{}
  \item{\code{A3}}{}
  \item{\code{B4}}{}
  \item{\code{B5_1a}}{}
  \item{\code{B5_1b}}{}
  \item{\code{B5_1c}}{}
  \item{\code{B5_2a}}{}
  \item{\code{B5_2b}}{}
  \item{\code{B5_2c}}{}
  \item{\code{B5_3a}}{}
  \item{\code{B5_3b}}{}
  \item{\code{B5_3c}}{}
  \item{\code{B5_4a}}{}
  \item{\code{B5_4b}}{}
  \item{\code{B5_4c}}{}
  \item{\code{B5_5a}}{}
  \item{\code{B5_5b}}{}
  \item{\code{B5_5c}}{}
  \item{\code{B5_6a}}{}
  \item{\code{B5_6b}}{}
  \item{\code{B5_6c}}{}
  \item{\code{B5_7a}}{}
  \item{\code{B5_7b}}{}
  \item{\code{B5_7c}}{}
  \item{\code{B5_8a}}{}
  \item{\code{B5_8b}}{}
  \item{\code{B5_8c}}{}
  \item{\code{B5_8_1a}}{}
  \item{\code{B5_8_1b}}{}
  \item{\code{B5_8_1c}}{}
  \item{\code{B6}}{}
  \item{\code{B7}}{}
  \item{\code{C8_1}}{}
  \item{\code{C8_2}}{}
  \item{\code{C8_3}}{}
  \item{\code{C8_4}}{}
  \item{\code{C9}}{}
  \item{\code{C11_1}}{}
  \item{\code{C11_2}}{}
  \item{\code{C11_3}}{}
  \item{\code{C11_4}}{}
  \item{\code{C11_5}}{}
  \item{\code{C11_6}}{}
  \item{\code{C11_7}}{}
  \item{\code{C11_8}}{}
  \item{\code{D12_1}}{}
  \item{\code{D12_2}}{}
  \item{\code{D13_1}}{}
  \item{\code{D13_2}}{}
  \item{\code{E14_1}}{}
  \item{\code{E14_2}}{}
  \item{\code{E14_3}}{}
  \item{\code{E14_4}}{}
  \item{\code{E14_5}}{}
  \item{\code{E14_6}}{}
  \item{\code{E14_7}}{}
  \item{\code{E14_8}}{}
  \item{\code{E14_9}}{}
  \item{\code{F15_1}}{}
  \item{\code{F15_2}}{}
  \item{\code{F15_3}}{}
  \item{\code{F15_4}}{}
  \item{\code{F15_5}}{}
  \item{\code{F15_6}}{}
  \item{\code{F15_7}}{}
  \item{\code{F15_8}}{}
  \item{\code{F15_9}}{}
  \item{\code{F15_10}}{}
  \item{\code{F15_11}}{}
  \item{\code{F15_12}}{}
  \item{\code{F15_13}}{}
  \item{\code{F15_14}}{}
  \item{\code{F15_15}}{}
  \item{\code{F15_16}}{}
  \item{\code{F15_17}}{}
  \item{\code{F15_18}}{}
  \item{\code{F15_19}}{}
  \item{\code{F15_20}}{}
  \item{\code{F15_21}}{}
  \item{\code{F15_22}}{}
  \item{\code{F15_23}}{}
  \item{\code{F15_24}}{}
  \item{\code{F15_25}}{}
  \item{\code{G16}}{}
  \item{\code{G17}}{}
  \item{\code{G18}}{}
  \item{\code{G18b}}{}
  \item{\code{G19}}{}
  \item{\code{G20}}{}
  \item{\code{G21}}{}
  \item{\code{G22_1}}{}
  \item{\code{G22_2}}{}
  \item{\code{G22_3}}{}
  \item{\code{G23_1}}{}
  \item{\code{G23_2}}{}
  \item{\code{G23_3}}{}
  \item{\code{G24}}{}
  \item{\code{G25_1}}{}
  \item{\code{G25_2}}{}
  \item{\code{G25_3}}{}
  \item{\code{G25_4}}{}
  \item{\code{G26_1}}{}
  \item{\code{G26_2}}{}
  \item{\code{G26_3}}{}
  \item{\code{G26_4}}{}
  \item{\code{G27_1}}{}
  \item{\code{G27_2}}{}
  \item{\code{G27_3}}{}
  \item{\code{G27_4}}{}
  \item{\code{G28_1}}{}
  \item{\code{G28_2}}{}
  \item{\code{G29_1}}{}
  \item{\code{G29_2}}{}
  \item{\code{G29_3}}{}
  \item{\code{G29_4}}{}
  \item{\code{G29_5}}{}
  \item{\code{G29_6}}{}
  \item{\code{G29_7}}{}
  \item{\code{G29_8}}{}
  \item{\code{G29_9}}{}
  \item{\code{G30}}{}
  \item{\code{G31}}{}
  \item{\code{G32_1}}{}
  \item{\code{G32_2}}{}
  \item{\code{G32_3}}{}
  \item{\code{G32_4}}{}
  \item{\code{G32_5}}{}
  \item{\code{H33}}{}
  \item{\code{H34}}{}
  \item{\code{H35}}{}
  \item{\code{H36}}{}
  \item{\code{H37}}{}
  \item{\code{H38}}{}
  \item{\code{H40}}{}
  \item{\code{H41}}{}
  \item{\code{H42}}{}
  \item{\code{H43}}{}
  \item{\code{H44}}{}
  \item{\code{H45}}{}
  \item{\code{H46_1}}{}
  \item{\code{H46_2}}{}
  \item{\code{H46_3}}{}
  \item{\code{H47_1}}{}
  \item{\code{H47_2}}{}
  \item{\code{H47_3}}{}
  \item{\code{H48_1}}{}
  \item{\code{H48_2}}{}
  \item{\code{H49}}{}
  \item{\code{H50}}{}
  \item{\code{H51}}{}
  \item{\code{H52}}{}
  \item{\code{H53_1}}{}
  \item{\code{H53_2}}{}
  \item{\code{H53_3}}{}
  \item{\code{H53_4}}{}
  \item{\code{H53_5}}{}
  \item{\code{H53_6}}{}
  \item{\code{H53_7}}{}
  \item{\code{H53_8}}{}
  \item{\code{H53_9}}{}
  \item{\code{I54}}{}
  \item{\code{I55_1a}}{}
  \item{\code{I55_1b}}{}
  \item{\code{I55_2a}}{}
  \item{\code{I55_2b}}{}
  \item{\code{I55_3a}}{}
  \item{\code{I55_3b}}{}
  \item{\code{I55_4a}}{}
  \item{\code{I55_4b}}{}
  \item{\code{I55_5a}}{}
  \item{\code{I55_5b}}{}
  \item{\code{I55_6a}}{}
  \item{\code{I55_6b}}{}
  \item{\code{I55_7a}}{}
  \item{\code{I55_7b}}{}
  \item{\code{I55_8a}}{}
  \item{\code{I55_8b}}{}
  \item{\code{I55_9a}}{}
  \item{\code{I55_9b}}{}
  \item{\code{I55_10a}}{}
  \item{\code{I55_10b}}{}
  \item{\code{I55_11a}}{}
  \item{\code{I55_11b}}{}
  \item{\code{I55_12a}}{}
  \item{\code{I55_12b}}{}
  \item{\code{I55_13a}}{}
  \item{\code{I55_13b}}{}
  \item{\code{I56_1}}{}
  \item{\code{I56_2}}{}
  \item{\code{J57}}{}
  \item{\code{J58}}{}
  \item{\code{J59}}{}
  \item{\code{K60}}{}
  \item{\code{K61}}{}
  \item{\code{K62}}{}
  \item{\code{K62b}}{}
  \item{\code{K63}}{}
  \item{\code{K63b}}{}
  \item{\code{L64}}{}
  \item{\code{L65_1}}{}
  \item{\code{L65_2}}{}
  \item{\code{L66}}{}
  \item{\code{L67_1}}{}
  \item{\code{L67_2}}{}
  \item{\code{L68}}{}
  \item{\code{L69_1}}{}
  \item{\code{L69_2}}{}
  \item{\code{L70}}{}
  \item{\code{L71_1}}{}
  \item{\code{L71_2}}{}
  \item{\code{L71_3}}{}
  \item{\code{L71_4}}{}
  \item{\code{L71_5}}{}
  \item{\code{L71_6}}{}
  \item{\code{L71_7}}{}
  \item{\code{L72_1}}{}
  \item{\code{L72_1a}}{}
  \item{\code{L72_2}}{}
  \item{\code{L72_3}}{}
  \item{\code{L72_4}}{}
  \item{\code{L72_5}}{}
  \item{\code{L72_6}}{}
  \item{\code{L72_7}}{}
  \item{\code{L72_7a}}{}
  \item{\code{L72_8}}{}
  \item{\code{L72_9}}{}
  \item{\code{L72_10}}{}
  \item{\code{L72_11}}{}
  \item{\code{L72_12}}{}
  \item{\code{L72_13}}{}
  \item{\code{L72_14}}{}
  \item{\code{L72_15}}{}
  \item{\code{L72_15a}}{}
  \item{\code{L72FZ}}{}
  \item{\code{L72CANTFZ}}{}
  \item{\code{L73}}{}
  \item{\code{L74}}{}
  \item{\code{M75}}{}
  \item{\code{M76_1}}{}
  \item{\code{M76_2}}{}
  \item{\code{M76_3}}{}
  \item{\code{M76_4}}{}
  \item{\code{M76_5}}{}
  \item{\code{N77}}{}
  \item{\code{MCALVIDA}}{}
  \item{\code{FACTORMENOR}}{}
  \item{\code{IMCm}}{}
  \item{\code{MSALMENT1}}{}
  \item{\code{MSALMENT2}}{}
  \item{\code{MSALMENT3}}{}
  \item{\code{MSALMENT4}}{}
  \item{\code{MSALMENT5}}{}
  \item{\code{CLASE_PR}}{}
}
}
\usage{
children_19
}
\description{
Variables of children survey of the ENSE 2017/19
}
\details{
DETAILS
}
\keyword{datasets}
