% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/children_19_info.R
\docType{data}
\name{children_19_info}
\alias{children_19_info}
\title{children_19_info}
\format{
A data frame with 267 rows and 6 variables:
\describe{
  \item{\code{variable_ine}}{variable name}
  \item{\code{longitud}}{field length}
  \item{\code{posicion_inicio}}{start position in fixed-width file}
  \item{\code{posicion_final}}{end position in fixed-width file}
  \item{\code{descripcion_del_campo}}{variable description}
  \item{\code{tipo_variable}}{thematic category of the variable}
}
}
\usage{
children_19_info
}
\description{
Design of the survey. Variable characteristics of children survey ENSE 2017/19.
}
\keyword{datasets}
