% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/children_19_labels.R
\docType{data}
\name{children_19_labels}
\alias{children_19_labels}
\title{children_19_labels}
\format{
A data frame with 1310 rows and 3 variables:
\describe{
  \item{\code{valores_ine}}{values of the variable}
  \item{\code{valores}}{labels}
  \item{\code{variable_ine}}{name of the variable}
}
}
\usage{
children_19_labels
}
\description{
Provides labels to the values of the children survey. ENSE 2017/19
}
\details{
DETAILS
}
\keyword{datasets}
