% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_region.R
\name{get_region}
\alias{get_region}
\title{Get variables per Spanish region (comunidad autonoma)}
\usage{
get_region(df, region)
}
\arguments{
\item{region}{Character string. Include the name of the region in lower case.
It accepts the following names: andalucia, aragon, asturias, baleares, canarias, cantabria,
castilla_leon, castilla la mancha, cataluna, comunidad valenciana, extremadura, galicia, madrid,}
}
\description{
This function provides information of the survey regarding particular regions.
}
