% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variable.R
\name{get_variable}
\alias{get_variable}
\title{Filter survey data by variable type}
\usage{
get_variable(df, info, type)
}
\arguments{
\item{df}{A data frame. One of the main survey datasets
(\code{adults_19}, \code{children_19}, \code{household_19}, etc.).}

\item{info}{A data frame. The corresponding \code{*_info} dataset
(\code{adults_19_info}, \code{children_19_info}, etc.). Must contain
columns \code{variable_ine} and \code{tipo_variable}.}

\item{type}{Character string. The variable category to select.
Use \code{unique(info$tipo_variable)} to list available types.}
}
\value{
A data frame with identifier columns plus all variables belonging
  to \code{type}.
}
\description{
Returns a subset of columns from a survey dataset that belong to a given
thematic category, as defined in the accompanying \code{*_info} dataset.
Identifier columns (\code{CCAA} and \code{IDENTHOGAR}) are always included
when present.
}
\examples{
# See available types
unique(adults_19_info$tipo_variable)

# Filter to health status variables
get_variable(adults_19, adults_19_info, "G. ESTADO DE SALUD")

}
