% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/house_06_labels.R
\docType{data}
\name{house_06_labels}
\alias{house_06_labels}
\title{house_06_labels}
\format{
A data frame with 611 rows and 2 variables:
\describe{
  \item{\code{variable_ine}}{variable name}
  \item{\code{valores_ine}}{coded value}
}
}
\usage{
house_06_labels
}
\description{
Value labels for the household survey ENSE 2006/07.
}
\keyword{datasets}
