% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/household_06.R
\docType{data}
\name{household_06}
\alias{household_06}
\title{household_06}
\format{
A data frame with 85044 rows and 84 variables:
\describe{
  \item{\code{NIVEL}}{}
  \item{\code{CCAA}}{}
  \item{\code{PROV}}{}
  \item{\code{ESTRATO}}{}
  \item{\code{SECCION}}{}
  \item{\code{SECC_DC}}{}
  \item{\code{NIDENTIF}}{}
  \item{\code{AÑO}}{}
  \item{\code{TRIM}}{}
  \item{\code{NORDEN}}{}
  \item{\code{P3_2}}{}
  \item{\code{A9_1}}{}
  \item{\code{A9_2}}{}
  \item{\code{A9_A2}}{}
  \item{\code{A9_M2}}{}
  \item{\code{C21}}{}
  \item{\code{C31}}{}
  \item{\code{EDAD}}{}
  \item{\code{SEXO}}{}
  \item{\code{A3}}{}
  \item{\code{A4}}{}
  \item{\code{A5}}{}
  \item{\code{A8}}{}
  \item{\code{A11}}{}
  \item{\code{A12}}{}
  \item{\code{A13}}{}
  \item{\code{B2}}{}
  \item{\code{B3}}{}
  \item{\code{B4}}{}
  \item{\code{B5}}{}
  \item{\code{B6}}{}
  \item{\code{C2_1}}{}
  \item{\code{C2_2}}{}
  \item{\code{C2_3}}{}
  \item{\code{C2_4}}{}
  \item{\code{C2_5}}{}
  \item{\code{C2_6}}{}
  \item{\code{C2_7}}{}
  \item{\code{C2_8}}{}
  \item{\code{C3}}{}
  \item{\code{C4}}{}
  \item{\code{C5}}{}
  \item{\code{C6}}{}
  \item{\code{C7}}{}
  \item{\code{C8}}{}
  \item{\code{C9}}{}
  \item{\code{C10}}{}
  \item{\code{C11}}{}
  \item{\code{C12}}{}
  \item{\code{C13}}{}
  \item{\code{C14}}{}
  \item{\code{C15}}{}
  \item{\code{C16}}{}
  \item{\code{C17}}{}
  \item{\code{C18}}{}
  \item{\code{C19}}{}
  \item{\code{D1_0_1}}{}
  \item{\code{D1_0_2}}{}
  \item{\code{D1}}{}
  \item{\code{D2}}{}
  \item{\code{D3_1}}{}
  \item{\code{D3_2}}{}
  \item{\code{D3_3}}{}
  \item{\code{D3_4}}{}
  \item{\code{D3_5}}{}
  \item{\code{D3_6}}{}
  \item{\code{D3_7}}{}
  \item{\code{D3_8}}{}
  \item{\code{E1_1}}{}
  \item{\code{E1_2}}{}
  \item{\code{E1_3}}{}
  \item{\code{E1_4}}{}
  \item{\code{E1_5}}{}
  \item{\code{E1_6}}{}
  \item{\code{E1_7}}{}
  \item{\code{E1_8}}{}
  \item{\code{E1_9}}{}
  \item{\code{E2}}{}
  \item{\code{E3}}{}
  \item{\code{FACTOR}}{}
  \item{\code{TMUNI}}{}
  \item{\code{CADULTO}}{}
  \item{\code{CNINO}}{}
  \item{\code{SPCLASE}}{}
}
}
\usage{
household_06
}
\description{
Household survey of the ENSE 2006/07.
}
\keyword{datasets}
