% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/household_06_info.R
\docType{data}
\name{household_06_info}
\alias{household_06_info}
\title{household_06_info}
\format{
A data frame with 84 rows and 7 variables:
\describe{
  \item{\code{campo}}{Name of the variable}
  \item{\code{longitud}}{Length of the variable. Number of possible characters it may take.}
  \item{\code{posicion_inicio}}{Initial position in the text file. Information for parsing microdata}
  \item{\code{posicion_final}}{Final position in the text file. Information for parsing microdata}
  \item{\code{descripcion_del_campo}}{Description of the variable (in Spanish)}
  \item{\code{valores_validos}}{Possible values the variable may take}
  \item{\code{tipo_variable}}{Type of variable. Only applicable to some surveys}
}
}
\usage{
household_06_info
}
\description{
Design of the survey. Variable characteristics of household survey ENSE 2006/07.
}
\details{
DETAILS
}
\keyword{datasets}
