% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/household_12.R
\docType{data}
\name{household_12}
\alias{household_12}
\title{household_12}
\format{
A data frame with 55806 rows and 66 variables:
\describe{
  \item{\code{CCAA}}{autonomous community code}
  \item{\code{IDENTHOGAR}}{household identifier}
  \item{\code{ESTRATO}}{sampling stratum}
  \item{\code{SEXO_i}}{sex of the reference person}
  \item{\code{EDAD_i}}{age of the reference person}
  \item{\code{NORDEN_Ai}}{order number of the adult reference person}
  \item{\code{NORDEN_Mi}}{order number of the minor reference person}
  \item{\code{NADULTOS}}{number of adults in the household}
  \item{\code{NMENORES}}{number of minors in the household}
  \item{\code{A7_1_i}}{}
  \item{\code{NORDEN_Pref}}{}
  \item{\code{A7_2a}}{}
  \item{\code{A7_2m}}{}
  \item{\code{A8_1_i}}{}
  \item{\code{A8_2_i}}{}
  \item{\code{NORINF}}{}
  \item{\code{A10_i}}{}
  \item{\code{A11_1_i}}{}
  \item{\code{A11_2_i}}{}
  \item{\code{A11_3_i}}{}
  \item{\code{A11_4_i}}{}
  \item{\code{A11_5_i}}{}
  \item{\code{A11_6_i}}{}
  \item{\code{A11_7_i}}{}
  \item{\code{A11_8_i}}{}
  \item{\code{A12}}{}
  \item{\code{B13}}{}
  \item{\code{B14}}{}
  \item{\code{B15_2}}{}
  \item{\code{B16_2}}{}
  \item{\code{B17}}{}
  \item{\code{B18}}{}
  \item{\code{B19a_2}}{}
  \item{\code{B19m_2}}{}
  \item{\code{B20a_2}}{}
  \item{\code{B20m_2}}{}
  \item{\code{B21a}}{}
  \item{\code{B21m}}{}
  \item{\code{C22}}{}
  \item{\code{C23}}{}
  \item{\code{C24_1}}{}
  \item{\code{C24_2}}{}
  \item{\code{C24_3}}{}
  \item{\code{C24_4}}{}
  \item{\code{C24_5}}{}
  \item{\code{C24_6}}{}
  \item{\code{C24_7}}{}
  \item{\code{C24_8}}{}
  \item{\code{C24_9}}{}
  \item{\code{FACTORHOGAR}}{sampling weight for the household}
  \item{\code{CADULTO}}{}
  \item{\code{CMENOR}}{}
  \item{\code{CLASE_PR}}{socioeconomic class of the reference person}
  \item{\code{D26_1}}{}
  \item{\code{D26_2}}{}
  \item{\code{D26_3}}{}
  \item{\code{D26_4}}{}
  \item{\code{D26_5}}{}
  \item{\code{D26_6}}{}
  \item{\code{D26_7}}{}
  \item{\code{D26_8}}{}
  \item{\code{D26_9}}{}
  \item{\code{D26_10}}{}
  \item{\code{D26_11}}{}
  \item{\code{D27}}{}
  \item{\code{D28}}{}
}
}
\usage{
household_12
}
\description{
Household survey of the ENSE 2011/12.
}
\keyword{datasets}
