% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/household_12_info.R
\docType{data}
\name{household_12_info}
\alias{household_12_info}
\title{household_12_info}
\format{
A data frame with 66 rows and 6 variables:
\describe{
  \item{\code{campo}}{variable name}
  \item{\code{longitud}}{field length}
  \item{\code{posicion_inicio}}{start position in fixed-width file}
  \item{\code{posicion_final}}{end position in fixed-width file}
  \item{\code{descripcion_del_campo}}{variable description}
  \item{\code{tipo_variable}}{thematic category of the variable}
}
}
\usage{
household_12_info
}
\description{
Design of the survey. Variable characteristics of household survey ENSE 2011/12.
}
\keyword{datasets}
