% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/household_12_labels.R
\docType{data}
\name{household_12_labels}
\alias{household_12_labels}
\title{household_12_labels}
\format{
A data frame with 269 rows and 3 variables:
\describe{
  \item{\code{valores_ine}}{coded value}
  \item{\code{valores}}{label}
  \item{\code{variable_ine}}{variable name}
}
}
\usage{
household_12_labels
}
\description{
Value labels for the household survey ENSE 2011/12.
}
\keyword{datasets}
