% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/household_19_labels.R
\docType{data}
\name{household_19_labels}
\alias{household_19_labels}
\title{household_19_labels}
\format{
A data frame with 312 rows and 3 variables:
\describe{
  \item{\code{valores_ine}}{values survey}
  \item{\code{valores}}{labels}
  \item{\code{variable_ine}}{name of the variable}
}
}
\usage{
household_19_labels
}
\description{
Provides labels to the values of the household survey. ENSE 17/19
}
\details{
DETAILS
}
\keyword{datasets}
