% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_data.R
\name{label_data}
\alias{label_data}
\title{Replace coded values with labels}
\usage{
label_data(df, labels, vars = NULL)
}
\arguments{
\item{df}{A data frame.}

\item{labels}{A labels data frame with columns `valores_ine`, `valores`, and `variable_ine`.}

\item{vars}{Optional character vector of variables to label.}
}
\value{
A data frame with selected variables labelled.
}
\description{
Replace coded values with labels
}
