% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_variable.R
\name{search_variable}
\alias{search_variable}
\title{Search variables by description}
\usage{
search_variable(info, pattern)
}
\arguments{
\item{info}{A variable metadata data frame.}

\item{pattern}{Text pattern to search for.}
}
\value{
A data frame with matching rows.
}
\description{
Search variables by description
}
