% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adults_06.R
\docType{data}
\name{adults_06}
\alias{adults_06}
\title{adults_06}
\format{
A data frame with 29478 rows and 619 variables:
\describe{
  \item{\code{NIVEL}}{}
  \item{\code{CCAA}}{}
  \item{\code{PROV *}}{}
  \item{\code{ESTRATO}}{}
  \item{\code{SECCION}}{}
  \item{\code{SECC_DC}}{}
  \item{\code{NIDENTIF}}{}
  \item{\code{A??O}}{}
  \item{\code{TRIM *}}{}
  \item{\code{NORDEN}}{}
  \item{\code{EDAD}}{}
  \item{\code{SEXO}}{}
  \item{\code{P3_3}}{}
  \item{\code{P3_4}}{}
  \item{\code{P3_5}}{}
  \item{\code{P4}}{}
  \item{\code{P5_0_1}}{}
  \item{\code{P5_0_2}}{}
  \item{\code{P5_2}}{}
  \item{\code{P5_3}}{}
  \item{\code{P5_4}}{}
  \item{\code{A1}}{}
  \item{\code{A2}}{}
  \item{\code{A3_1}}{}
  \item{\code{A3_2}}{}
  \item{\code{A4_1}}{}
  \item{\code{A4_2}}{}
  \item{\code{A5}}{}
  \item{\code{A6_1}}{}
  \item{\code{A6_2}}{}
  \item{\code{A7_1}}{}
  \item{\code{A7_2}}{}
  \item{\code{A8}}{}
  \item{\code{A9_1}}{}
  \item{\code{A9_2}}{}
  \item{\code{A10_1}}{}
  \item{\code{A10_2}}{}
  \item{\code{A11_1}}{}
  \item{\code{A11_2}}{}
  \item{\code{A12_1}}{}
  \item{\code{A12_2}}{}
  \item{\code{A13}}{}
  \item{\code{B14}}{}
  \item{\code{B15_1A}}{}
  \item{\code{B15_1B}}{}
  \item{\code{B15_1C}}{}
  \item{\code{B15_1D}}{}
  \item{\code{B15_2A}}{}
  \item{\code{B15_2B}}{}
  \item{\code{B15_2C}}{}
  \item{\code{B15_2D}}{}
  \item{\code{B15_3A}}{}
  \item{\code{B15_3B}}{}
  \item{\code{B15_3C}}{}
  \item{\code{B15_3D}}{}
  \item{\code{B15_4A}}{}
  \item{\code{B15_4B}}{}
  \item{\code{B15_4C}}{}
  \item{\code{B15_4D}}{}
  \item{\code{B15_5A}}{}
  \item{\code{B15_5B}}{}
  \item{\code{B15_5C}}{}
  \item{\code{B15_5D}}{}
  \item{\code{B15_6A}}{}
  \item{\code{B15_6B}}{}
  \item{\code{B15_6C}}{}
  \item{\code{B15_6D}}{}
  \item{\code{B15_7A}}{}
  \item{\code{B15_7B}}{}
  \item{\code{B15_7C}}{}
  \item{\code{B15_7D}}{}
  \item{\code{B15_8A}}{}
  \item{\code{B15_8B}}{}
  \item{\code{B15_8C}}{}
  \item{\code{B15_8D}}{}
  \item{\code{B15_9A}}{}
  \item{\code{B15_9B}}{}
  \item{\code{B15_9C}}{}
  \item{\code{B15_9D}}{}
  \item{\code{B15_10A}}{}
  \item{\code{B15_10B}}{}
  \item{\code{B15_10C}}{}
  \item{\code{B15_10D}}{}
  \item{\code{B15_11A}}{}
  \item{\code{B15_11B}}{}
  \item{\code{B15_11C}}{}
  \item{\code{B15_11D}}{}
  \item{\code{B15_12A}}{}
  \item{\code{B15_12B}}{}
  \item{\code{B15_12C}}{}
  \item{\code{B15_12D}}{}
  \item{\code{B15_13A}}{}
  \item{\code{B15_13B}}{}
  \item{\code{B15_13C}}{}
  \item{\code{B15_13D}}{}
  \item{\code{B15_14A}}{}
  \item{\code{B15_14B}}{}
  \item{\code{B15_14C}}{}
  \item{\code{B15_14D}}{}
  \item{\code{B15_15A}}{}
  \item{\code{B15_15B}}{}
  \item{\code{B15_15C}}{}
  \item{\code{B15_15D}}{}
  \item{\code{B15_16A}}{}
  \item{\code{B15_16B}}{}
  \item{\code{B15_16C}}{}
  \item{\code{B15_16D}}{}
  \item{\code{B15_17A}}{}
  \item{\code{B15_17B}}{}
  \item{\code{B15_17C}}{}
  \item{\code{B15_17D}}{}
  \item{\code{B15_18A}}{}
  \item{\code{B15_18B}}{}
  \item{\code{B15_18C}}{}
  \item{\code{B15_18D}}{}
  \item{\code{B15_19A}}{}
  \item{\code{B15_19B}}{}
  \item{\code{B15_19C}}{}
  \item{\code{B15_19D}}{}
  \item{\code{B15_20A}}{}
  \item{\code{B15_20B}}{}
  \item{\code{B15_20C}}{}
  \item{\code{B15_20D}}{}
  \item{\code{B15_21A}}{}
  \item{\code{B15_21B}}{}
  \item{\code{B15_21C}}{}
  \item{\code{B15_21D}}{}
  \item{\code{B15_22A}}{}
  \item{\code{B15_22B}}{}
  \item{\code{B15_22C}}{}
  \item{\code{B15_22D}}{}
  \item{\code{B15_23A}}{}
  \item{\code{B15_23B}}{}
  \item{\code{B15_23C}}{}
  \item{\code{B15_23D}}{}
  \item{\code{B15_24A}}{}
  \item{\code{B15_24B}}{}
  \item{\code{B15_24C}}{}
  \item{\code{B15_24D}}{}
  \item{\code{B15_25A}}{}
  \item{\code{B15_25B}}{}
  \item{\code{B15_25C}}{}
  \item{\code{B15_25D}}{}
  \item{\code{B15_26A}}{}
  \item{\code{B15_26B}}{}
  \item{\code{B15_26C}}{}
  \item{\code{B15_26D}}{}
  \item{\code{B15_27A}}{}
  \item{\code{B15_27B}}{}
  \item{\code{B15_27C}}{}
  \item{\code{B15_27D}}{}
  \item{\code{B15_28A}}{}
  \item{\code{B15_28LI}}{}
  \item{\code{B15_28B}}{}
  \item{\code{B15_28C}}{}
  \item{\code{B15_28D}}{}
  \item{\code{B15_29A}}{}
  \item{\code{B15_29LI}}{}
  \item{\code{B15_29B}}{}
  \item{\code{B15_29C}}{}
  \item{\code{B15_29D}}{}
  \item{\code{B16}}{}
  \item{\code{C17}}{}
  \item{\code{C18_1A}}{}
  \item{\code{C18_1B}}{}
  \item{\code{C18_2A}}{}
  \item{\code{C18_2B}}{}
  \item{\code{C18_3A}}{}
  \item{\code{C18_3B}}{}
  \item{\code{C18_4A}}{}
  \item{\code{C18_4B}}{}
  \item{\code{C18_5A}}{}
  \item{\code{C18_5B}}{}
  \item{\code{C18_6A}}{}
  \item{\code{C18_6B}}{}
  \item{\code{C18_7A}}{}
  \item{\code{C18_7B}}{}
  \item{\code{C18_8A}}{}
  \item{\code{C18_8B}}{}
  \item{\code{C19}}{}
  \item{\code{C20}}{}
  \item{\code{C21_1}}{}
  \item{\code{C21_2}}{}
  \item{\code{C21_3}}{}
  \item{\code{C21_4}}{}
  \item{\code{C21_5}}{}
  \item{\code{D22_1}}{}
  \item{\code{D22_2}}{}
  \item{\code{D23_1}}{}
  \item{\code{D23_2}}{}
  \item{\code{D24_1A}}{}
  \item{\code{D24_1B}}{}
  \item{\code{D24_2A}}{}
  \item{\code{D24_2B}}{}
  \item{\code{D24_3A}}{}
  \item{\code{D24_3B}}{}
  \item{\code{D24_4A}}{}
  \item{\code{D24_4B}}{}
  \item{\code{D24_5A}}{}
  \item{\code{D24_5B}}{}
  \item{\code{D24_6A}}{}
  \item{\code{D24_6B}}{}
  \item{\code{D24_7A}}{}
  \item{\code{D24_7B}}{}
  \item{\code{D24_8A}}{}
  \item{\code{D24_8B}}{}
  \item{\code{D24_9A}}{}
  \item{\code{D24_9B}}{}
  \item{\code{D24_10A}}{}
  \item{\code{D24_10B}}{}
  \item{\code{D24_11A}}{}
  \item{\code{D24_11B}}{}
  \item{\code{D24_12A}}{}
  \item{\code{D24_12B}}{}
  \item{\code{D24_13A}}{}
  \item{\code{D24_13B}}{}
  \item{\code{D24_14A}}{}
  \item{\code{D24_14B}}{}
  \item{\code{D24_15A}}{}
  \item{\code{D24_15B}}{}
  \item{\code{D24_16A}}{}
  \item{\code{D24_16B}}{}
  \item{\code{D24_17A}}{}
  \item{\code{D24_17B}}{}
  \item{\code{D24_18A}}{}
  \item{\code{D24_18B}}{}
  \item{\code{D24_19A}}{}
  \item{\code{D24_19B}}{}
  \item{\code{E25_1A}}{}
  \item{\code{E25_1B}}{}
  \item{\code{E25_2A}}{}
  \item{\code{E25_2B}}{}
  \item{\code{E25_3A}}{}
  \item{\code{E25_3B}}{}
  \item{\code{E25_4A}}{}
  \item{\code{E25_4B}}{}
  \item{\code{E25_5A}}{}
  \item{\code{E25_5B}}{}
  \item{\code{E25_6A}}{}
  \item{\code{E25_6B}}{}
  \item{\code{E25_7A}}{}
  \item{\code{E25_7B}}{}
  \item{\code{E25_8A}}{}
  \item{\code{E25_8B}}{}
  \item{\code{E25_9A}}{}
  \item{\code{E25_9B}}{}
  \item{\code{E25_10A}}{}
  \item{\code{E25_10B}}{}
  \item{\code{E25_11A}}{}
  \item{\code{E25_11B}}{}
  \item{\code{E25_12A}}{}
  \item{\code{E25_12B}}{}
  \item{\code{E25_13A}}{}
  \item{\code{E25_13B}}{}
  \item{\code{E25_14A}}{}
  \item{\code{E25_14B}}{}
  \item{\code{E25_15A}}{}
  \item{\code{E25_15B}}{}
  \item{\code{E25_16A}}{}
  \item{\code{E25_16B}}{}
  \item{\code{E25_17A}}{}
  \item{\code{E25_17B}}{}
  \item{\code{E25_18A}}{}
  \item{\code{E25_18B}}{}
  \item{\code{E25_19A}}{}
  \item{\code{E25_19B}}{}
  \item{\code{E25_20A}}{}
  \item{\code{E25_20B}}{}
  \item{\code{E25_21A}}{}
  \item{\code{E25_22A}}{}
  \item{\code{F26_1}}{}
  \item{\code{F26_2}}{}
  \item{\code{F26_3}}{}
  \item{\code{F26_4}}{}
  \item{\code{F26_5}}{}
  \item{\code{F26_6}}{}
  \item{\code{F26_7}}{}
  \item{\code{F26_8}}{}
  \item{\code{F26_9}}{}
  \item{\code{F26_10}}{}
  \item{\code{F26_11}}{}
  \item{\code{F26_12}}{}
  \item{\code{F27_0}}{}
  \item{\code{F27}}{}
  \item{\code{F28}}{}
  \item{\code{G1_29_1}}{}
  \item{\code{G1_29_2}}{}
  \item{\code{G1_29_3}}{}
  \item{\code{G1_30}}{}
  \item{\code{G1_31}}{}
  \item{\code{G1_32}}{}
  \item{\code{G1_33}}{}
  \item{\code{G1_34}}{}
  \item{\code{G1_35_1}}{}
  \item{\code{G1_35_2}}{}
  \item{\code{G1_35_3}}{}
  \item{\code{G1_36_1}}{}
  \item{\code{G1_36_2}}{}
  \item{\code{G1_36_3}}{}
  \item{\code{G1_37}}{}
  \item{\code{G1_38}}{}
  \item{\code{G1_39_1A}}{}
  \item{\code{G1_39_1B}}{}
  \item{\code{G1_39_1C}}{}
  \item{\code{G1_39_2A}}{}
  \item{\code{G1_39_2B}}{}
  \item{\code{G1_39_2C}}{}
  \item{\code{G1_39_3A}}{}
  \item{\code{G1_39_3B}}{}
  \item{\code{G1_39_3C}}{}
  \item{\code{G1_39_4A}}{}
  \item{\code{G1_39_4B}}{}
  \item{\code{G1_39_4C}}{}
  \item{\code{G1_39_5A}}{}
  \item{\code{G1_39_5B}}{}
  \item{\code{G1_39_5C}}{}
  \item{\code{G1_39_6A}}{}
  \item{\code{G1_39_6B}}{}
  \item{\code{G1_39_6C}}{}
  \item{\code{G1_40_1A}}{}
  \item{\code{G1_40_1B}}{}
  \item{\code{G1_40_2A}}{}
  \item{\code{G1_40_2B}}{}
  \item{\code{G1_40_3A}}{}
  \item{\code{G1_40_3B}}{}
  \item{\code{G2_41}}{}
  \item{\code{G2_42}}{}
  \item{\code{G3_43_1}}{}
  \item{\code{G3_43_2}}{}
  \item{\code{G3_43_3}}{}
  \item{\code{G3_43_4}}{}
  \item{\code{G3_43_5}}{}
  \item{\code{G3_43_6}}{}
  \item{\code{G3_43_7}}{}
  \item{\code{G3_43_8}}{}
  \item{\code{G4_44_1}}{}
  \item{\code{G4_44_2}}{}
  \item{\code{G4_44_3}}{}
  \item{\code{G4_44_4}}{}
  \item{\code{G4_45_1}}{}
  \item{\code{G4_45_2}}{}
  \item{\code{G4_45_3}}{}
  \item{\code{G4_45_4}}{}
  \item{\code{G4_45_5}}{}
  \item{\code{G4_45_6}}{}
  \item{\code{G4_45_7}}{}
  \item{\code{G4_45_8}}{}
  \item{\code{G4_45_9}}{}
  \item{\code{G4_46}}{}
  \item{\code{G4_47_1}}{}
  \item{\code{G4_47_2}}{}
  \item{\code{G4_47_3}}{}
  \item{\code{G4_47_4}}{}
  \item{\code{G4_47_5}}{}
  \item{\code{G4_47_6}}{}
  \item{\code{G4_47_7}}{}
  \item{\code{G4_47_8}}{}
  \item{\code{G5_48}}{}
  \item{\code{G5_49}}{}
  \item{\code{G5_50}}{}
  \item{\code{G5_51}}{}
  \item{\code{G5_52}}{}
  \item{\code{G5_53}}{}
  \item{\code{G5_54}}{}
  \item{\code{G5_55}}{}
  \item{\code{G5_56}}{}
  \item{\code{G5_57}}{}
  \item{\code{G5_58}}{}
  \item{\code{G5_59}}{}
  \item{\code{G6_60}}{}
  \item{\code{G6_61}}{}
  \item{\code{G6_62_1}}{}
  \item{\code{G6_62_2}}{}
  \item{\code{G6_62_3}}{}
  \item{\code{G6_63_1}}{}
  \item{\code{G6_63_2}}{}
  \item{\code{G6_63_3}}{}
  \item{\code{G6_64_1}}{}
  \item{\code{G6_64_2}}{}
  \item{\code{G6_65}}{}
  \item{\code{G6_66}}{}
  \item{\code{H1_67}}{}
  \item{\code{H1_68_1}}{}
  \item{\code{H1_68_2}}{}
  \item{\code{H1_68_3}}{}
  \item{\code{H1_69}}{}
  \item{\code{H1_70}}{}
  \item{\code{H1_71}}{}
  \item{\code{H1_72_1}}{}
  \item{\code{H1_72_2}}{}
  \item{\code{H1_72_3}}{}
  \item{\code{H1_73}}{}
  \item{\code{H1_74}}{}
  \item{\code{H1_75}}{}
  \item{\code{H1_76_1}}{}
  \item{\code{H1_76_2}}{}
  \item{\code{H1_77_1}}{}
  \item{\code{H1_77_2}}{}
  \item{\code{H1_77_3}}{}
  \item{\code{H1_77_4}}{}
  \item{\code{H1_77_5}}{}
  \item{\code{H1_77_6}}{}
  \item{\code{H1_77_7}}{}
  \item{\code{H1_77LIT}}{}
  \item{\code{H1_78_1}}{}
  \item{\code{H1_78_2}}{}
  \item{\code{H1_78_3}}{}
  \item{\code{H1_79_1}}{}
  \item{\code{H1_79_2}}{}
  \item{\code{H1_80_1}}{}
  \item{\code{H1_80_2}}{}
  \item{\code{H2_81}}{}
  \item{\code{H2_82}}{}
  \item{\code{H2_83}}{}
  \item{\code{H2_84_1A}}{}
  \item{\code{H2_84_1B}}{}
  \item{\code{H2_84_1C}}{}
  \item{\code{H2_84_2A}}{}
  \item{\code{H2_84_2B}}{}
  \item{\code{H2_84_2C}}{}
  \item{\code{H2_84_3A}}{}
  \item{\code{H2_84_3B}}{}
  \item{\code{H2_84_3C}}{}
  \item{\code{H2_84_4A}}{}
  \item{\code{H2_84_4B}}{}
  \item{\code{H2_84_4C}}{}
  \item{\code{H2_84_5A}}{}
  \item{\code{H2_84_5B}}{}
  \item{\code{H2_84_5C}}{}
  \item{\code{H2_84_6A}}{}
  \item{\code{H2_84_6B}}{}
  \item{\code{H2_84_6C}}{}
  \item{\code{H2_85_1}}{}
  \item{\code{H2_85_2}}{}
  \item{\code{H2_85_3}}{}
  \item{\code{H2_85_4}}{}
  \item{\code{H2_85_5}}{}
  \item{\code{H2_85_6}}{}
  \item{\code{H2_86}}{}
  \item{\code{H3_87}}{}
  \item{\code{H3_88}}{}
  \item{\code{H3_89_1}}{}
  \item{\code{H3_89_2}}{}
  \item{\code{H3_89_3}}{}
  \item{\code{H3_90}}{}
  \item{\code{H3_91}}{}
  \item{\code{H3_92}}{}
  \item{\code{H3_93}}{}
  \item{\code{H3_94_1}}{}
  \item{\code{H3_94_2}}{}
  \item{\code{H3_94_3}}{}
  \item{\code{H4_95_1}}{}
  \item{\code{H4_95_2}}{}
  \item{\code{H4_95_3}}{}
  \item{\code{H4_95_4}}{}
  \item{\code{H4_95_5}}{}
  \item{\code{H4_95_6}}{}
  \item{\code{H4_96_1}}{}
  \item{\code{H4_96_2}}{}
  \item{\code{H4_96_3}}{}
  \item{\code{H4_96_4}}{}
  \item{\code{H4_96_5}}{}
  \item{\code{H4_96_6}}{}
  \item{\code{H4_96_7}}{}
  \item{\code{H4_96_8}}{}
  \item{\code{H4_96_9}}{}
  \item{\code{H4_96_10}}{}
  \item{\code{H4_96_11}}{}
  \item{\code{H4_96_12}}{}
  \item{\code{H4_97}}{}
  \item{\code{H4_98}}{}
  \item{\code{I1_99}}{}
  \item{\code{I1_100}}{}
  \item{\code{I1_101}}{}
  \item{\code{I1_102}}{}
  \item{\code{I1_103}}{}
  \item{\code{I1_104}}{}
  \item{\code{I2_105}}{}
  \item{\code{I3_106_0}}{}
  \item{\code{I3_106}}{}
  \item{\code{I3_107}}{}
  \item{\code{I3_108}}{}
  \item{\code{I3_109}}{}
  \item{\code{I3_110}}{}
  \item{\code{I3_111}}{}
  \item{\code{I3_112_1}}{}
  \item{\code{I3_112_2}}{}
  \item{\code{I3_112_3}}{}
  \item{\code{I3_112_4}}{}
  \item{\code{I3_112_5}}{}
  \item{\code{I3_112_6}}{}
  \item{\code{I3_112_7}}{}
  \item{\code{I3_112_8}}{}
  \item{\code{I3_113}}{}
  \item{\code{I3_114}}{}
  \item{\code{I3_115}}{}
  \item{\code{I3_116}}{}
  \item{\code{I3_117}}{}
  \item{\code{I4_118_1}}{}
  \item{\code{I4_118_2}}{}
  \item{\code{I4_119_1}}{}
  \item{\code{I4_119_2}}{}
  \item{\code{I4_120_1}}{}
  \item{\code{I4_120_2}}{}
  \item{\code{I4_121_1}}{}
  \item{\code{I4_121_2}}{}
  \item{\code{J122_0}}{}
  \item{\code{J122}}{}
  \item{\code{J123_0}}{}
  \item{\code{J123}}{}
  \item{\code{J124}}{}
  \item{\code{J125_1}}{}
  \item{\code{J125_2}}{}
  \item{\code{J125_3}}{}
  \item{\code{J125_4}}{}
  \item{\code{J125_5}}{}
  \item{\code{J126}}{}
  \item{\code{J127_1A}}{}
  \item{\code{J127_1B}}{}
  \item{\code{J127_1C}}{}
  \item{\code{J127_1D}}{}
  \item{\code{J127_1E}}{}
  \item{\code{J127_1F}}{}
  \item{\code{J127_2A}}{}
  \item{\code{J127_2B}}{}
  \item{\code{J127_2C}}{}
  \item{\code{J127_2D}}{}
  \item{\code{J127_2E}}{}
  \item{\code{J127_2F}}{}
  \item{\code{J127_3A}}{}
  \item{\code{J127_3B}}{}
  \item{\code{J127_3C}}{}
  \item{\code{J127_3D}}{}
  \item{\code{J127_3E}}{}
  \item{\code{J127_3F}}{}
  \item{\code{J127_4A}}{}
  \item{\code{J127_4B}}{}
  \item{\code{J127_4C}}{}
  \item{\code{J127_4D}}{}
  \item{\code{J127_4E}}{}
  \item{\code{J127_4F}}{}
  \item{\code{J127_5A}}{}
  \item{\code{J127_5B}}{}
  \item{\code{J127_5C}}{}
  \item{\code{J127_5D}}{}
  \item{\code{J127_5E}}{}
  \item{\code{J127_5F}}{}
  \item{\code{J127_6A}}{}
  \item{\code{J127_6B}}{}
  \item{\code{J127_6C}}{}
  \item{\code{J127_6D}}{}
  \item{\code{J127_6E}}{}
  \item{\code{J127_6F}}{}
  \item{\code{L1_128}}{}
  \item{\code{L1_129}}{}
  \item{\code{L1_130}}{}
  \item{\code{L2_131}}{}
  \item{\code{L2_132}}{}
  \item{\code{L2_133}}{}
  \item{\code{L2_134}}{}
  \item{\code{M135_1}}{}
  \item{\code{M135_2}}{}
  \item{\code{M135_3}}{}
  \item{\code{M135_4}}{}
  \item{\code{M135_5}}{}
  \item{\code{M135_6}}{}
  \item{\code{M135_7}}{}
  \item{\code{M135_8}}{}
  \item{\code{M135_9}}{}
  \item{\code{M135_10}}{}
  \item{\code{M135_11}}{}
  \item{\code{M136_1}}{}
  \item{\code{M136_2}}{}
  \item{\code{M136_3}}{}
  \item{\code{M136_4}}{}
  \item{\code{M136_5}}{}
  \item{\code{N137}}{}
  \item{\code{N138}}{}
  \item{\code{N139}}{}
  \item{\code{N140_1}}{}
  \item{\code{N140_2}}{}
  \item{\code{N140_3}}{}
  \item{\code{N140_4}}{}
  \item{\code{N140_5}}{}
  \item{\code{N140_6}}{}
  \item{\code{N140_7}}{}
  \item{\code{N140_8}}{}
  \item{\code{N140_9}}{}
  \item{\code{N140_10}}{}
  \item{\code{N140_11}}{}
  \item{\code{N140_12}}{}
  \item{\code{N140_13}}{}
  \item{\code{N140_14}}{}
  \item{\code{N140_15}}{}
  \item{\code{N140_16}}{}
  \item{\code{N140_17}}{}
  \item{\code{N140_18}}{}
  \item{\code{N140_19}}{}
  \item{\code{N140_20}}{}
  \item{\code{N140_21}}{}
  \item{\code{N140_22}}{}
  \item{\code{N140_23}}{}
  \item{\code{N140_24}}{}
  \item{\code{N140_25}}{}
  \item{\code{N140_26}}{}
  \item{\code{N140_27}}{}
  \item{\code{FACTOR}}{}
  \item{\code{TMUNI}}{}
  \item{\code{SPCLASE}}{}
  \item{\code{ASCLASE}}{}
  \item{\code{ASRELACTI}}{}
  \item{\code{ASSITUPRO}}{}
  \item{\code{ASSALMENT}}{}
  \item{\code{ASDEPFUN1}}{}
  \item{\code{ASDEPFUN2}}{}
  \item{\code{ASDEPFUN3}}{}
  \item{\code{ASAPSOCIAL}}{}
  \item{\code{ASFUNFAMI}}{}
  \item{\code{ASIMC}}{}
}
}
\usage{
adults_06
}
\description{
Variables of adult survey of the ENSE 2006/07
}
\details{
DETAILS
}
\keyword{datasets}
