% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adults_12.R
\docType{data}
\name{adults_12}
\alias{adults_12}
\title{adults_12}
\format{
A data frame with 21007 rows and 577 variables:
\describe{
  \item{\code{CCAA}}{}
  \item{\code{IDENTHOGAR}}{}
  \item{\code{ESTRATO}}{}
  \item{\code{A7_2a}}{}
  \item{\code{SEXOa}}{}
  \item{\code{EDADa}}{}
  \item{\code{PROXY_0}}{}
  \item{\code{PROXY_1}}{}
  \item{\code{PROXY_2}}{}
  \item{\code{PROXY_2b}}{}
  \item{\code{PROXY_3a}}{}
  \item{\code{PROXY_4}}{}
  \item{\code{PROXY_5}}{}
  \item{\code{E1_1}}{}
  \item{\code{PAISEXT}}{}
  \item{\code{E2_1a}}{}
  \item{\code{E2_1b}}{}
  \item{\code{E2_1c}}{}
  \item{\code{E2_1d}}{}
  \item{\code{NACEXT}}{}
  \item{\code{E3}}{}
  \item{\code{E4}}{}
  \item{\code{F5}}{}
  \item{\code{F6}}{}
  \item{\code{F7_2}}{}
  \item{\code{F8_2}}{}
  \item{\code{F9}}{}
  \item{\code{F10}}{}
  \item{\code{F11}}{}
  \item{\code{F12}}{}
  \item{\code{F13a}}{}
  \item{\code{F13m}}{}
  \item{\code{F14}}{}
  \item{\code{F15}}{}
  \item{\code{F16a_2}}{}
  \item{\code{F16m_2}}{}
  \item{\code{F17a_2}}{}
  \item{\code{F17m_2}}{}
  \item{\code{F18}}{}
  \item{\code{G19}}{}
  \item{\code{G20}}{}
  \item{\code{G21a_1}}{}
  \item{\code{G21b_1}}{}
  \item{\code{G21c_1}}{}
  \item{\code{G21a_2}}{}
  \item{\code{G21b_2}}{}
  \item{\code{G21c_2}}{}
  \item{\code{G21a_3}}{}
  \item{\code{G21b_3}}{}
  \item{\code{G21c_3}}{}
  \item{\code{G21a_4}}{}
  \item{\code{G21b_4}}{}
  \item{\code{G21c_4}}{}
  \item{\code{G21a_5}}{}
  \item{\code{G21b_5}}{}
  \item{\code{G21c_5}}{}
  \item{\code{G21a_6}}{}
  \item{\code{G21b_6}}{}
  \item{\code{G21c_6}}{}
  \item{\code{G21a_7}}{}
  \item{\code{G21b_7}}{}
  \item{\code{G21c_7}}{}
  \item{\code{G21a_8}}{}
  \item{\code{G21b_8}}{}
  \item{\code{G21c_8}}{}
  \item{\code{G21a_9}}{}
  \item{\code{G21b_9}}{}
  \item{\code{G21c_9}}{}
  \item{\code{G21a_10}}{}
  \item{\code{G21b_10}}{}
  \item{\code{G21c_10}}{}
  \item{\code{G21a_11}}{}
  \item{\code{G21b_11}}{}
  \item{\code{G21c_11}}{}
  \item{\code{G21a_12}}{}
  \item{\code{G21b_12}}{}
  \item{\code{G21c_12}}{}
  \item{\code{G21a_13}}{}
  \item{\code{G21b_13}}{}
  \item{\code{G21c_13}}{}
  \item{\code{G21a_14}}{}
  \item{\code{G21b_14}}{}
  \item{\code{G21c_14}}{}
  \item{\code{G21a_15}}{}
  \item{\code{G21b_15}}{}
  \item{\code{G21c_15}}{}
  \item{\code{G21a_16}}{}
  \item{\code{G21b_16}}{}
  \item{\code{G21c_16}}{}
  \item{\code{G21a_17}}{}
  \item{\code{G21b_17}}{}
  \item{\code{G21c_17}}{}
  \item{\code{G21a_18}}{}
  \item{\code{G21b_18}}{}
  \item{\code{G21c_18}}{}
  \item{\code{G21a_19}}{}
  \item{\code{G21b_19}}{}
  \item{\code{G21c_19}}{}
  \item{\code{G21a_20}}{}
  \item{\code{G21b_20}}{}
  \item{\code{G21c_20}}{}
  \item{\code{G21a_21}}{}
  \item{\code{G21b_21}}{}
  \item{\code{G21c_21}}{}
  \item{\code{G21a_22}}{}
  \item{\code{G21b_22}}{}
  \item{\code{G21c_22}}{}
  \item{\code{G21a_23}}{}
  \item{\code{G21b_23}}{}
  \item{\code{G21c_23}}{}
  \item{\code{G21a_24}}{}
  \item{\code{G21b_24}}{}
  \item{\code{G21c_24}}{}
  \item{\code{G21a_25}}{}
  \item{\code{G21b_25}}{}
  \item{\code{G21c_25}}{}
  \item{\code{G21a_26}}{}
  \item{\code{G21b_26}}{}
  \item{\code{G21c_26}}{}
  \item{\code{G21a_27}}{}
  \item{\code{G21b_27}}{}
  \item{\code{G21c_27}}{}
  \item{\code{G21a_28}}{}
  \item{\code{G21b_28}}{}
  \item{\code{G21c_28}}{}
  \item{\code{G21a_29}}{}
  \item{\code{G21b_29}}{}
  \item{\code{G21c_29}}{}
  \item{\code{G21a_30}}{}
  \item{\code{G21b_30}}{}
  \item{\code{G21c_30}}{}
  \item{\code{G21_31}}{}
  \item{\code{G21_31_1b}}{}
  \item{\code{G21_31_1c}}{}
  \item{\code{G21_31_2b}}{}
  \item{\code{G21_31_2c}}{}
  \item{\code{G21_31_3b}}{}
  \item{\code{G21_31_3c}}{}
  \item{\code{H22}}{}
  \item{\code{H23}}{}
  \item{\code{H24}}{}
  \item{\code{H25_1}}{}
  \item{\code{H25_2}}{}
  \item{\code{H25_3}}{}
  \item{\code{H25_4}}{}
  \item{\code{H25_5}}{}
  \item{\code{H25_6}}{}
  \item{\code{H25_7}}{}
  \item{\code{H25_8}}{}
  \item{\code{I26.1}}{}
  \item{\code{I26.2}}{}
  \item{\code{I26.3}}{}
  \item{\code{I26.4}}{}
  \item{\code{I26.5}}{}
  \item{\code{I26.6}}{}
  \item{\code{J27_1}}{}
  \item{\code{J27_2}}{}
  \item{\code{J28_1}}{}
  \item{\code{J28_2}}{}
  \item{\code{K29}}{}
  \item{\code{K30}}{}
  \item{\code{K31_1}}{}
  \item{\code{K31_2}}{}
  \item{\code{K31_3}}{}
  \item{\code{K31_4}}{}
  \item{\code{K31_5}}{}
  \item{\code{K31_6}}{}
  \item{\code{K31_7}}{}
  \item{\code{K31_8}}{}
  \item{\code{K31_9}}{}
  \item{\code{K31_10}}{}
  \item{\code{K31_11}}{}
  \item{\code{K31_12}}{}
  \item{\code{K31_13}}{}
  \item{\code{K31_14}}{}
  \item{\code{K31_15}}{}
  \item{\code{K31_16}}{}
  \item{\code{K31_17}}{}
  \item{\code{K31_18}}{}
  \item{\code{K31_19}}{}
  \item{\code{K31_20}}{}
  \item{\code{K31_21}}{}
  \item{\code{K31_22}}{}
  \item{\code{K31_23}}{}
  \item{\code{K31_24}}{}
  \item{\code{K31_25}}{}
  \item{\code{K31_26}}{}
  \item{\code{K31_27}}{}
  \item{\code{L32}}{}
  \item{\code{L33}}{}
  \item{\code{L34}}{}
  \item{\code{L35}}{}
  \item{\code{L36}}{}
  \item{\code{L37}}{}
  \item{\code{L38}}{}
  \item{\code{L39}}{}
  \item{\code{M40_1}}{}
  \item{\code{M40_2}}{}
  \item{\code{M40_3}}{}
  \item{\code{M40_4}}{}
  \item{\code{M40_5}}{}
  \item{\code{M40_6}}{}
  \item{\code{M40_7}}{}
  \item{\code{M40_8}}{}
  \item{\code{M40_9}}{}
  \item{\code{M40_10}}{}
  \item{\code{M40_11}}{}
  \item{\code{M40_12}}{}
  \item{\code{M41}}{}
  \item{\code{M42}}{}
  \item{\code{N43_1}}{}
  \item{\code{N43_2}}{}
  \item{\code{N43_3}}{}
  \item{\code{N44_1}}{}
  \item{\code{N44_2}}{}
  \item{\code{N45_1}}{}
  \item{\code{N45_2}}{}
  \item{\code{N46}}{}
  \item{\code{N47}}{}
  \item{\code{N48}}{}
  \item{\code{N49_1}}{}
  \item{\code{N49_2}}{}
  \item{\code{N49_3}}{}
  \item{\code{N50_1}}{}
  \item{\code{N50_2}}{}
  \item{\code{N50_3}}{}
  \item{\code{N51}}{}
  \item{\code{N52_1}}{}
  \item{\code{N52_2}}{}
  \item{\code{N52_3}}{}
  \item{\code{N52_4}}{}
  \item{\code{N52_5}}{}
  \item{\code{N53_1}}{}
  \item{\code{N53_2}}{}
  \item{\code{N53_3}}{}
  \item{\code{N53_4}}{}
  \item{\code{N54_1}}{}
  \item{\code{N54_2}}{}
  \item{\code{N54_3}}{}
  \item{\code{N54_4}}{}
  \item{\code{N55_1}}{}
  \item{\code{N55_2}}{}
  \item{\code{N55_3}}{}
  \item{\code{N55_4}}{}
  \item{\code{N55_5}}{}
  \item{\code{N56_1}}{}
  \item{\code{N56_2}}{}
  \item{\code{N57_1}}{}
  \item{\code{N57_2}}{}
  \item{\code{N57_3}}{}
  \item{\code{N57_4}}{}
  \item{\code{N57_5}}{}
  \item{\code{N57_6}}{}
  \item{\code{N57_7}}{}
  \item{\code{N57_8}}{}
  \item{\code{N57_9}}{}
  \item{\code{N57_10}}{}
  \item{\code{N58}}{}
  \item{\code{N59_1}}{}
  \item{\code{N59_2}}{}
  \item{\code{N59_3}}{}
  \item{\code{N59_4}}{}
  \item{\code{N59_5}}{}
  \item{\code{N59_6}}{}
  \item{\code{N59_7}}{}
  \item{\code{N59_8}}{}
  \item{\code{O60}}{}
  \item{\code{O61}}{}
  \item{\code{O62}}{}
  \item{\code{O63}}{}
  \item{\code{O64}}{}
  \item{\code{O65}}{}
  \item{\code{O66}}{}
  \item{\code{O67}}{}
  \item{\code{O68}}{}
  \item{\code{O69}}{}
  \item{\code{O70}}{}
  \item{\code{O71}}{}
  \item{\code{O72}}{}
  \item{\code{O73_1}}{}
  \item{\code{O73_2}}{}
  \item{\code{O73_3}}{}
  \item{\code{O74_1}}{}
  \item{\code{O74_2}}{}
  \item{\code{O74_3}}{}
  \item{\code{O75_1}}{}
  \item{\code{O75_2}}{}
  \item{\code{O76}}{}
  \item{\code{O77}}{}
  \item{\code{O78}}{}
  \item{\code{O79}}{}
  \item{\code{O80_1}}{}
  \item{\code{O80_2}}{}
  \item{\code{O80_3}}{}
  \item{\code{O80_4}}{}
  \item{\code{O80_5}}{}
  \item{\code{O80_6}}{}
  \item{\code{O80_7}}{}
  \item{\code{O80_8}}{}
  \item{\code{O80_9}}{}
  \item{\code{P81_a}}{}
  \item{\code{P81_b}}{}
  \item{\code{P82_1a}}{}
  \item{\code{P82_1b}}{}
  \item{\code{P82_2a}}{}
  \item{\code{P82_2b}}{}
  \item{\code{P82_3a}}{}
  \item{\code{P82_3b}}{}
  \item{\code{P82_4a}}{}
  \item{\code{P82_4b}}{}
  \item{\code{P82_5a}}{}
  \item{\code{P82_5b}}{}
  \item{\code{P82_6a}}{}
  \item{\code{P82_6b}}{}
  \item{\code{P82_7a}}{}
  \item{\code{P82_7b}}{}
  \item{\code{P82_8a}}{}
  \item{\code{P82_8b}}{}
  \item{\code{P82_9a}}{}
  \item{\code{P82_9b}}{}
  \item{\code{P82_10a}}{}
  \item{\code{P82_10b}}{}
  \item{\code{P82_11a}}{}
  \item{\code{P82_11b}}{}
  \item{\code{P82_12a}}{}
  \item{\code{P82_12b}}{}
  \item{\code{P82_13a}}{}
  \item{\code{P82_13b}}{}
  \item{\code{P82_14a}}{}
  \item{\code{P82_14b}}{}
  \item{\code{P82_15a}}{}
  \item{\code{P82_15b}}{}
  \item{\code{P82_16a}}{}
  \item{\code{P82_16b}}{}
  \item{\code{P82_17a}}{}
  \item{\code{P82_17b}}{}
  \item{\code{P82_18a}}{}
  \item{\code{P82_18b}}{}
  \item{\code{P82_19a}}{}
  \item{\code{P82_19b}}{}
  \item{\code{P82_20a}}{}
  \item{\code{P82_20b}}{}
  \item{\code{P82_21a}}{}
  \item{\code{P82_21b}}{}
  \item{\code{P83_1}}{}
  \item{\code{P83_2}}{}
  \item{\code{Q84}}{}
  \item{\code{Q85}}{}
  \item{\code{Q86}}{}
  \item{\code{Q87}}{}
  \item{\code{Q88}}{}
  \item{\code{Q89}}{}
  \item{\code{Q90}}{}
  \item{\code{Q91}}{}
  \item{\code{Q92}}{}
  \item{\code{Q93}}{}
  \item{\code{Q94}}{}
  \item{\code{Q95}}{}
  \item{\code{Q96}}{}
  \item{\code{Q97}}{}
  \item{\code{Q98}}{}
  \item{\code{Q99}}{}
  \item{\code{Q100}}{}
  \item{\code{Q101}}{}
  \item{\code{R102}}{}
  \item{\code{R103}}{}
  \item{\code{R104}}{}
  \item{\code{S105}}{}
  \item{\code{S106a_1}}{}
  \item{\code{S106a_2}}{}
  \item{\code{S106a_3}}{}
  \item{\code{S106b_1}}{}
  \item{\code{S106b_2}}{}
  \item{\code{S106b_3}}{}
  \item{\code{S107}}{}
  \item{\code{S108}}{}
  \item{\code{S109}}{}
  \item{\code{S110}}{}
  \item{\code{S111}}{}
  \item{\code{S112_1}}{}
  \item{\code{S112_2}}{}
  \item{\code{S113_1}}{}
  \item{\code{S113_2}}{}
  \item{\code{S113_3}}{}
  \item{\code{S113_4}}{}
  \item{\code{S113_5}}{}
  \item{\code{S113a_6}}{}
  \item{\code{S113_7}}{}
  \item{\code{S113_8}}{}
  \item{\code{S114}}{}
  \item{\code{S115_1}}{}
  \item{\code{S115_2}}{}
  \item{\code{S115_3}}{}
  \item{\code{S115_4}}{}
  \item{\code{S115_5}}{}
  \item{\code{S115_6}}{}
  \item{\code{S115_7}}{}
  \item{\code{S115_8}}{}
  \item{\code{S116a_1}}{}
  \item{\code{S116a_2}}{}
  \item{\code{S116a_3}}{}
  \item{\code{S116b_1}}{}
  \item{\code{S116b_2}}{}
  \item{\code{S116b_3}}{}
  \item{\code{S117}}{}
  \item{\code{S118}}{}
  \item{\code{S119}}{}
  \item{\code{T120}}{}
  \item{\code{T121}}{}
  \item{\code{T122}}{}
  \item{\code{T123}}{}
  \item{\code{T124Cer}}{}
  \item{\code{T124Cer_1}}{}
  \item{\code{T124Cer_2}}{}
  \item{\code{T124Cer_3}}{}
  \item{\code{T124Cer_4}}{}
  \item{\code{T124Cer_5}}{}
  \item{\code{T124Cer_6}}{}
  \item{\code{T124Cer_7}}{}
  \item{\code{T124Vin}}{}
  \item{\code{T124Vin_1}}{}
  \item{\code{T124Vin_2}}{}
  \item{\code{T124Vin_3}}{}
  \item{\code{T124Vin_4}}{}
  \item{\code{T124Vin_5}}{}
  \item{\code{T124Vin_6}}{}
  \item{\code{T124Vin_7}}{}
  \item{\code{T124Vermut}}{}
  \item{\code{T124Vermut_1}}{}
  \item{\code{T124Vermut_2}}{}
  \item{\code{T124Vermut_3}}{}
  \item{\code{T124Vermut_4}}{}
  \item{\code{T124Vermut_5}}{}
  \item{\code{T124Vermut_6}}{}
  \item{\code{T124Vermut_7}}{}
  \item{\code{T124Lic}}{}
  \item{\code{T124Lic_1}}{}
  \item{\code{T124Lic_2}}{}
  \item{\code{T124Lic_3}}{}
  \item{\code{T124Lic_4}}{}
  \item{\code{T124Lic_5}}{}
  \item{\code{T124Lic_6}}{}
  \item{\code{T124Lic_7}}{}
  \item{\code{T124Comb}}{}
  \item{\code{T124Comb_1}}{}
  \item{\code{T124Comb_2}}{}
  \item{\code{T124Comb_3}}{}
  \item{\code{T124Comb_4}}{}
  \item{\code{T124Comb_5}}{}
  \item{\code{T124Comb_6}}{}
  \item{\code{T124Comb_7}}{}
  \item{\code{T124Sidra}}{}
  \item{\code{T124Sidra_1}}{}
  \item{\code{T124Sidra_2}}{}
  \item{\code{T124Sidra_3}}{}
  \item{\code{T124Sidra_4}}{}
  \item{\code{T124Sidra_5}}{}
  \item{\code{T124Sidra_6}}{}
  \item{\code{T124Sidra_7}}{}
  \item{\code{T125}}{}
  \item{\code{T126a}}{}
  \item{\code{T126b}}{}
  \item{\code{U127}}{}
  \item{\code{U128}}{}
  \item{\code{U129}}{}
  \item{\code{U130_1}}{}
  \item{\code{U130_2a}}{}
  \item{\code{U130_2b}}{}
  \item{\code{U130_3}}{}
  \item{\code{U130_4a}}{}
  \item{\code{U130_4b}}{}
  \item{\code{U130_5}}{}
  \item{\code{U130_6a}}{}
  \item{\code{U130_6b}}{}
  \item{\code{U130_7a}}{}
  \item{\code{U130_7b}}{}
  \item{\code{V131}}{}
  \item{\code{V132_1}}{}
  \item{\code{V132_2}}{}
  \item{\code{V132_3}}{}
  \item{\code{V132_4}}{}
  \item{\code{V132_5}}{}
  \item{\code{V132_6}}{}
  \item{\code{V132_7}}{}
  \item{\code{V133_1}}{}
  \item{\code{V133_1a}}{}
  \item{\code{V133_2}}{}
  \item{\code{V133_3}}{}
  \item{\code{V133_4}}{}
  \item{\code{V133_5}}{}
  \item{\code{V133_6}}{}
  \item{\code{V133_7}}{}
  \item{\code{V133_7a}}{}
  \item{\code{V133_8}}{}
  \item{\code{V133_9}}{}
  \item{\code{V133_10}}{}
  \item{\code{V133_11}}{}
  \item{\code{V133_12}}{}
  \item{\code{V133_13}}{}
  \item{\code{V133_14}}{}
  \item{\code{V133_15}}{}
  \item{\code{V133_15a}}{}
  \item{\code{V134}}{}
  \item{\code{V135}}{}
  \item{\code{V136}}{}
  \item{\code{W137}}{}
  \item{\code{W138_1}}{}
  \item{\code{W138_2}}{}
  \item{\code{W138_3}}{}
  \item{\code{W138_4}}{}
  \item{\code{W138_5}}{}
  \item{\code{Y139_1}}{}
  \item{\code{Y139_2}}{}
  \item{\code{Y139_3}}{}
  \item{\code{Y139_4}}{}
  \item{\code{Y139_5}}{}
  \item{\code{Y139_6}}{}
  \item{\code{Y139_7}}{}
  \item{\code{Y139_8}}{}
  \item{\code{Y139_9}}{}
  \item{\code{Y139_10}}{}
  \item{\code{Y139_11}}{}
  \item{\code{Z140}}{}
  \item{\code{Z140_a1}}{}
  \item{\code{Z140_a2}}{}
  \item{\code{Z140_a3}}{}
  \item{\code{Z140_a4}}{}
  \item{\code{Z140_a5}}{}
  \item{\code{Z140_a6}}{}
  \item{\code{Z140_a7}}{}
  \item{\code{Z140_a8}}{}
  \item{\code{Z140_a9}}{}
  \item{\code{Z140_a10}}{}
  \item{\code{Z140_a11}}{}
  \item{\code{Z140_a12}}{}
  \item{\code{Z140_a13}}{}
  \item{\code{Z140_a14}}{}
  \item{\code{Z140_a15}}{}
  \item{\code{Z140_m1}}{}
  \item{\code{Z140_m2}}{}
  \item{\code{Z140_m3}}{}
  \item{\code{Z140_m4}}{}
  \item{\code{Z140_m5}}{}
  \item{\code{Z140_m6}}{}
  \item{\code{Z140_m7}}{}
  \item{\code{Z140_m8}}{}
  \item{\code{Z140_m9}}{}
  \item{\code{Z140_m10}}{}
  \item{\code{Z140_m11}}{}
  \item{\code{Z140_m12}}{}
  \item{\code{Z140_m13}}{}
  \item{\code{Z140_m14}}{}
  \item{\code{Z140_m15}}{}
  \item{\code{Z141}}{}
  \item{\code{Z141_1}}{}
  \item{\code{Z141_2}}{}
  \item{\code{Z142}}{}
  \item{\code{Z142_1}}{}
  \item{\code{Z143_1}}{}
  \item{\code{Z143_2}}{}
  \item{\code{Z144}}{}
  \item{\code{Z144_1}}{}
  \item{\code{Z144_2}}{}
  \item{\code{Z145}}{}
  \item{\code{Z145_1}}{}
  \item{\code{Z145_2}}{}
  \item{\code{FACTORADULTO}}{}
  \item{\code{CLASE_PR}}{}
  \item{\code{IMCa}}{}
  \item{\code{DEPFUNC1}}{}
  \item{\code{DEPFUNC2}}{}
  \item{\code{DEPFUNC3}}{}
  \item{\code{SALUDMENTAL}}{}
  \item{\code{APSOCIAL}}{}
  \item{\code{CMD1}}{}
  \item{\code{CMD2}}{}
  \item{\code{CMD3}}{}
}
}
\usage{
adults_12
}
\description{
Variables of adult survey of the ENSE 2011/12
}
\details{
DETAILS
}
\keyword{datasets}
