% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enseResp-package.R
\docType{package}
\name{enseResp-package}
\alias{enseResp}
\alias{enseResp-package}
\title{enseResp: Information on the Spanish Health Survey}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
