% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enseResp-package.R
\docType{package}
\name{enseResp-package}
\alias{enseResp}
\alias{enseResp-package}
\title{enseResp: Information on the Spanish Health Survey}
\description{
This package provides information in the Spanish Health Survey. It contains three
   samples: adults, family and kids.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/edugonzaloalmorox/enseResp}
  \item Report bugs at \url{https://github.com/edugonzaloalmorox/enseResp/issues}
}

}
\keyword{internal}
