% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/household_19_info.R
\docType{data}
\name{household_19_info}
\alias{household_19_info}
\title{household_19_info}
\format{A data frame with 58 rows and 6 variables:
\describe{
  \item{\code{variable_ine}}{character COLUMN_DESCRIPTION}
  \item{\code{longitud}}{double COLUMN_DESCRIPTION}
  \item{\code{posicion_inicio}}{double COLUMN_DESCRIPTION}
  \item{\code{posicion_final}}{double COLUMN_DESCRIPTION}
  \item{\code{descripcion_del_campo}}{character COLUMN_DESCRIPTION}
  \item{\code{tipo_variable}}{character COLUMN_DESCRIPTION}
}}
\usage{
household_19_info
}
\description{
Design of the survey. Variable characteristics of children survey.
}
\details{
DETAILS
}
\keyword{datasets}
