% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/household_19_info.R
\docType{data}
\name{household_19_info}
\alias{household_19_info}
\title{household_19_info}
\format{
A data frame with 58 rows and 6 variables:
\describe{
  \item{\code{campo}}{Name of the variable}
  \item{\code{longitud}}{Length of the variable. Number of possible characters it may take.}
  \item{\code{posicion_inicio}}{Initial position in the text file. Information for parsing microdata}
  \item{\code{posicion_final}}{Final position in the text file. Information for parsing microdata}
  \item{\code{descripcion_del_campo}}{Description of the variable (in Spanish)}
  \item{\code{valores_validos}}{Possible values the variable may take}
}
}
\usage{
household_19_info
}
\description{
Design of the survey. Variable characteristics of children survey ENSE 2017/19.
}
\details{
DETAILS
}
\keyword{datasets}
